### 2024-11-18 #673 @ryanzomorrodi
`sasquatch`: Use SAS, R, and Quarto together
Submitting Author Name: Ryan Zomorrodi
Submitting Author Github Handle: @ryanzomorrodi
Other Package Authors Github handles: (comma separated, delete if none)
Repository:  https://github.com/ryanzomorrodi/sasquatch
Version submitted: 0.0.0.9004
Submission type: Standard
Version accepted: TBD

---

```
Package: sasquatch
Title: Use 'SAS', R, and 'quarto' Together
Version: 0.0.0.9004
    person("Ryan", "Zomorrodi", , "rzomor2@uic.edu", role = c("aut", "cre", "cph"),
Description: Use R and 'SAS' within reproducible multilingual 'quarto'
    documents. Run 'SAS' code blocks interactively, send data back and forth
    between 'SAS' and R, and render 'SAS' output within quarto documents. 'SAS'
    connections are established through a combination of 'SASPy' and
    'reticulate'.
SystemRequirements:
    python (>= 3.0.0),
    SASPy,
    java (>= 7) required for IOM access method
    htmlwidgets,
    reticulate,
    rstudioapi
```

`sasquatch` allows users to create remote SAS sessions, work with them interactively, pass tables between R and SAS, and render quarto documents.

The target audience would be anyone looking to combine SAS in R within a single workflow. Using quarto, analysts can have one single reproducable report that combines SAS and R (and python potentially).

`saquatch` works similarly to packages like [`sasr`](https://sassoftware.github.io/saspy/) or [`configSAS`](https://github.com/baselr/configSAS). `sasr`, while similar to `sasquatch`, does not include interactive SAS functionality or a `knitr` engine. On the other hand, `configSAS` includes a `knitr` engine, but no interactive SAS  functionality. `configSAS` `knitr` output also does not include syntax highlighting and nested SAS output interferes with the styles of the rest of the document.

N/A

Locally, `pkgcheck` is able to pass all tests except coverage. I do not currently have any coverage over the `knitr` engine because I am struggling to think about how I can test that effectively. My assumption that the automated `pkgcheck` testing will fail due to the need for `python`,  python package `SASPy`, `java`,  and a SAS License. (I have set up a CI that has all of these set up within the `sasquatch` repo).

- [ ] contains a vignette with examples of its essential functions and uses. (this would need to be rendered locally)
