+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   setting_up: setting_up.html
-last_built: 2024-11-18T20:49Z
+last_built: 2024-11-19T21:47Z
 urls:
   reference: https://ryanzomorrodi.github.io/sasquatch/reference
   article: https://ryanzomorrodi.github.io/sasquatch/articles
