+++ tests/testthat/test-knitr.R
@@ -1,4 +1,6 @@
 test_that("multiplication works", {
+  skip_if_offline()
+
   sas_use_quarto()
 
   files <- c("sasquatch.qmd", "sas.xml")
