### 2024-11-20 2594a59
Adding suggests curl for skip_if_offline
