+++ DESCRIPTION
@@ -25,7 +25,8 @@ Imports:
 URL: https://ryanzomorrodi.github.io/sasquatch/, https://github.com/ryanzomorrodi/sasr
 Suggests: 
     rmarkdown,
-    testthat (>= 3.0.0)
+    testthat (>= 3.0.0),
+    curl
 VignetteBuilder: knitr
 Config/testthat/edition: 3
 BugReports: https://github.com/ryanzomorrodi/sasr/issues
