+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   setting_up: setting_up.html
-last_built: 2024-11-19T22:04Z
+last_built: 2024-11-20T00:52Z
 urls:
   reference: https://ryanzomorrodi.github.io/sasquatch/reference
   article: https://ryanzomorrodi.github.io/sasquatch/articles
