### 2024-11-20 933f7b5
Remove syntax highlighting and sas_use_quarto
