### 2024-11-20 c596349
Update README and setting up to reflect simpler quarto
