+++ inst/sasquatch.qmd
@@ -1,18 +1,0 @@
----
-format:
-  html:
-    syntax-definitions: 
-    - sas.xml
-include-in-header:
-  text: <script src = 'https://cdn.jsdelivr.net/npm/@iframe-resizer/parent@5.3.2'></script>
-engine: knitr
----
-
-```{r, include = FALSE}
-library(sasquatch)
-sas_connect()
-```
-
-```{sas}
-
-```
