+++ _pkgdown.yml
@@ -18,7 +18,6 @@ reference:
   - sas_run_string
   - sas_run_selected
   - sas_run_file
-  - sas_use_quarto
   - sas_engine
 
 - title: R <-> SAS
