+++ README.Rmd
@@ -71,14 +71,14 @@ sas_connect()
 
 Now, you should be able to run SAS code blocks in RStudio like any other.
 
-![Executing a SAS chunk](man/figures/run_sas_chunk.gif)
+![](man/figures/run_sas_chunk.gif)
 
 #### Sending output to viewer
 
 If you want to send the SAS output to the viewer, you can utilize the 
 `sas_run_selected()` addin with a custom shortcut.
 
-![Running SAS Selection](man/figures/run_sas_selected.gif)
+![](man/figures/run_sas_selected.gif)
 
 #### Converting tables
 
@@ -94,7 +94,7 @@ cars <- sas_to_r("cars", libref = "sashelp")
 And of course, render beautiful quarto documents in the same style you would
 expect from SAS with the `sas_engine()`.
 
-![Rendered Quarto Document](man/figures/rendered_quarto.png)
+![](man/figures/rendered_quarto.png)
 
 ## Similar packages
 
