### 2024-11-22 dc28555
Remove background color from quarto iframe
