### 2024-11-23 a12e6a7
Add installation and configuration functions
