+++ R/install.R
@@ -1,0 +1,174 @@
+#' Install SASPy package
+#' 
+#' @description
+#' Installs the SASPy package within a virtual environment.
+#' 
+#' @param envname Name of virtual environment to install SASPy within. 
+#' @param extra_packages Additional packages to install.
+#' @param restart_session Restart session? 
+#' 
+#' @return No return value.
+#' 
+#' @export
+#' 
+#' @examples
+#' \dontrun{
+#' install_saspy()
+#' }
+install_saspy <- function(
+  envname = "r-saspy",
+  extra_packages = NULL,
+  restart_session = TRUE
+) {
+  packages <- unique(c("wheel", "saspy", "pandas"), extra_packages)
+  
+  if (reticulate::virtualenv_exists(envname)) {
+    reticulate::virtualenv_remove(envname = envname, confirm = FALSE)
+  }
+  reticulate::py_install(
+    packages = packages,
+    envname = envname
+  )
+    
+  if (restart_session && rstudioapi::hasFun("restartSession")) {
+    rstudioapi::restartSession()
+  }
+
+  invisible(NULL)
+}
+
+#' Configure SASPy package
+#' 
+#' @description
+#' Adds `sascfg_personal.py` and `authinfo` files and prefills relevant info
+#' according to a specified template.
+#' 
+#' @param template Default template to base configuration files off of.
+#' 
+#' @return No return value.
+#' 
+#' @export
+#' 
+#' @examples
+#' \dontrun{
+#' config_saspy()
+#' }
+configure_saspy <- function(
+  template = c("none", "oda")
+) {
+  python <- reticulate::py_discover_config("saspy", "r-saspy")
+  saspy_path <- python$required_module_path
+  sascfg_personal_path <- paste0(saspy_path, "/sascfg_personal.py")
+
+  config_name <- "config_name"
+  config <- list()
+
+  if (template[1] == "oda") {
+    config_name <- "oda"
+    oda_username <- readline("Enter your ODA username: ")
+    oda_password <- readline("Enter your ODA password: ")
+    authinfo <- paste("oda user", oda_username, "password", oda_password)
+
+    if(.Platform$OS.type == "windows") {
+      authinfo_path <- paste0(get_home_dir(), "/_authinfo")
+    } else {
+      authinfo_path <- paste0(get_home_dir(), "/.authinfo")
+    }
+
+    if (file.exists(authinfo_path)) {
+      overwrite_auth = menu(
+        c("Yes", "No"), 
+        title = "The authinfo file already exists and you're about to overwrite it.\nDo you want to proceed anyway?"
+      )
+      if (overwrite_auth == 0) {
+        stop("The authinfo file already exists.")
+      }
+    }
+    if (!exists("overwrite_auth") || overwrite_auth == 1) {
+      cat("Writing to", authinfo_path, "\n")
+      cat(authinfo, file = authinfo_path)
+    }
+
+    servers <- list(
+      "['odaws01-usw2.oda.sas.com','odaws02-usw2.oda.sas.com','odaws03-usw2.oda.sas.com','odaws04-usw2.oda.sas.com']",
+      "['odaws01-usw2-2.oda.sas.com','odaws02-usw2-2.oda.sas.com']",
+      "['odaws01-euw1.oda.sas.com','odaws02-euw1.oda.sas.com']",
+      "['odaws01-apse1.oda.sas.com','odaws02-apse1.oda.sas.com']",
+      "['odaws01-apse1-2.oda.sas.com','odaws02-apse1-2.oda.sas.com']"
+    )
+    server_num <- menu(
[…]
