+++ man/install_saspy.Rd
@@ -1,0 +1,30 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/install.R
+\name{install_saspy}
+\alias{install_saspy}
+\title{Install SASPy package}
+\usage{
+install_saspy(
+  envname = "r-saspy",
+  extra_packages = NULL,
+  restart_session = TRUE
+)
+}
+\arguments{
+\item{envname}{Name of virtual environment to install SASPy within.}
+
+\item{extra_packages}{Additional packages to install.}
+
+\item{restart_session}{Restart session?}
+}
+\value{
+No return value.
+}
+\description{
+Installs the SASPy package within a virtual environment.
+}
+\examples{
+\dontrun{
+install_saspy()
+}
+}
