+++ NAMESPACE
@@ -1,5 +1,7 @@
 # Generated by roxygen2: do not edit by hand
 
+export(configure_saspy)
+export(install_saspy)
 export(r_to_sas)
 export(sas_connect)
 export(sas_disconnect)
