### 2024-11-23 ebd6324
Simplify the setup vignette
