+++ vignettes/setting_up.Rmd
@@ -1,198 +1,0 @@
----
-title: "Setting Up"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Setting Up}
-  %\VignetteEngine{knitr::rmarkdown}
-  %\VignetteEncoding{UTF-8}
----
-
-```{r, include = FALSE}
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
-```
-
-`sasquatch` works by utilizing the [`SASPy`](https://sassoftware.github.io/SASPy/) python package, similar to packages like [`sasr`](https://github.com/insightsengineering/sasr) or [configSAS](https://github.com/baselr/configSAS). This means everything we do to connect R and SAS, needs to go through the translation layer which is `SASPy`.
-
-Setting up can be difficult, but this guide is here to help you! Keep in mind there are many different ways to connect to SAS and different operating systems will have different steps. The following vignette will only address connecting to SAS On Demand for Academics (ODA) servers. If you are looking to utilize other SAS clients, check out the [`SASPy` configuration documentation](https://sassoftware.github.io/SASPy/configuration.html#). Although it is quite lengthy, it will provide you with all the necessary steps to connect to any SAS client from any machine.
-
-## Signing up for SAS ODA
-
-SAS ODA is free SAS client for professors, students, and independent learners. Create an account at <https://welcome.oda.sas.com/>.
-
-Once you have set up your account, log in and note the ODA server (in the picture below United States 2) and your username (under the email in the profile dropdown). We will need these for later.
-
-![](images/sas_oda.png)
-
-## Installing Python
-
-If you do not already have Python already installed, [install Python](https://www.python.org/downloads/) from the Python Software Foundation website.
-
-## Installing Java
-
-If you do not already have Java installed, install [Java](https://www.java.com/en/download/). Note the path that you install java in.
-
-If you do have Java installed, but do not remember where you installed it, then...
-
-In windows, it is usually installed in `C:\Program Files\Java\jdk{version number}`, but you can double check by opening up a terminal and typing in:
-
-```         
-where java
-```
-
-In mac or linux, it is usually installed in `/usr/bin/java`, but you can double check by opening up a terminal and typing in:
-
-```         
-which java
-```
-
-Note the Java path for later.
-
-## Setting up `SASPy`
-
-Note we are going install the `SASPy` package in our global python install. It's often a better practice to create a virtual environment, conda environment, or use one of the many other ways to isolate package installs. This isn't necessarily a worry if you only plan to use SAS and R together.
-
-In windows, open up a command prompt type in:
-
-```         
-py -m pip install wheel
-py -m pip install SASPy
-py -m pip install pandas
-py -m site --user-site
-```
-
-In linux or max, open up a terminal and type in:
-
-```         
-python3 -m pip install wheel
-python3 -m pip install SASPy
-python3 -m pip install pandas
-python3 -m site --user-site
-```
-
-This will install the required packages for `SASPy` and then print out the path where they were installed. Navigate to this path and find the `SASPy` folder. This is where the `SASPy` package is stored. Create a new file within the SASPy package called `sascfg_personal.py`.
-
[…]
