### 2024-11-23 6114622
Add install.packages("pak") to README
