+++ .github/workflows/test-coverage.yaml
@@ -38,18 +38,7 @@ jobs:
 
       - name: setup r-reticulate venv
         shell: Rscript {0}
-        run: |
-
-          path_to_python <- reticulate::virtualenv_create(
-            envname = "r-reticulate",
-            python = Sys.which("python"), # placed on PATH by the setup-python action
-            packages = c(
-              "wheel", "saspy", "pandas"
-            )
-          )
-
-          writeLines(sprintf("RETICULATE_PYTHON=%s", path_to_python),
-                     Sys.getenv("GITHUB_ENV"))
+        run: sasquatch::install_saspy()
 
       - name: setup saspy
         env:
@@ -59,7 +48,7 @@ jobs:
           oda_username = Sys.getenv("SAS_ODA_USERNAME")
           oda_password = Sys.getenv("SAS_ODA_PASSWORD")
           cat(paste("oda user", oda_username,  "password", oda_password), file = "~/.authinfo")
-          saspy_path <- paste0(gsub("Location: ", "", system(paste(reticulate::virtualenv_python("r-reticulate"), "-m pip show saspy"), intern = TRUE)[8]), "/saspy")
+          saspy_path <- paste0(gsub("Location: ", "", system(paste(reticulate::virtualenv_python("r-saspy"), "-m pip show saspy"), intern = TRUE)[8]), "/saspy")
           cat("SAS_config_names = ['oda']
           
           oda = {
