### 2024-11-23 #673 @emilyriederer
Hi @ryanzomorrodi ! Thanks for your submission.

I see your `README.md` draws contrast with `sasR` and `configSAS`. Could you please also comment how this compares to the "native" `sas` [language engine](https://bookdown.org/yihui/rmarkdown-cookbook/eng-sas.html) for R Markdown? Is it largely the interactive capabilities you mention? Or the ability to use SODA versus an on-computer executable?

(Transparently, I'm not a SAS user, so I lack some background context on this space.)
