### 2024-11-23 #673 @ryanzomorrodi
Hi @emilyriederer ,

I have never used the native `sas` engine. I saw the mention of it within the RMarkdown cookbook before, but frankly I can't find any other mention of someone using it in the wild. Unfortunately, I don't have access to a computer with a SAS executable installed to test it out.

One package I forgot to mention is [`SASmarkdown`](https://cran.r-project.org/web/packages/SASmarkdown/index.html). It also relies on a local SAS executable so I haven't been able to test it out, but like you said, it doesn't have any interactive features.

One of the advantages in relying on the SASPy Python package is SASPy can connect to both local and remote SAS clients. The configuration is a bit tedious, but at least for SODA, `configure_saspy()` can do it for you. I was hoping that if others use it, they could share how they set up their configurations and then additional templates can be added.
