+++ README.Rmd
@@ -21,14 +21,14 @@ knitr::opts_chunk$set(
 coverage](https://codecov.io/gh/ryanzomorrodi/sasquatch/branch/main/graph/badge.svg)](https://app.codecov.io/gh/ryanzomorrodi/sasquatch?branch=main)
 <!-- badges: end -->
 
-Use SAS, R, and Quarto together
+Use SAS, R, and Quarto Together
 
 `sasquatch` allows you to combine the power of R, SAS, and quarto together
 to create reproducible multilingual reports. `sasquatch` can run SAS code blocks
 interactively, send data back and forth between SAS and R, and render SAS HTML 
 output within quarto documents.
 
-`sasquatch` relies on the [`SASPy`](https://sassoftware.github.io/saspy/) Python package. But if you
+`sasquatch` relies on the [`SASPy`](https://sassoftware.github.io/saspy/) Python package. But if you...
 
 - Don't have `SASPy` already installed, or  
 - Don't have a SAS License  
