### 2024-12-04 #673 @emilyriederer
Hi @ryanzomorrodi - I'm happy to share that we think `sasquatch` is in scope. Given the necessary external dependencies like SAS it may take a bit longer to get our checks run and to identify the right editor, but more to come ASAP.

(Also, what an absolutely perfect name...)
