### 2024-12-04 #673 @ryanzomorrodi
That seems like a very reasonable approach, I will try it out. Thank you
