### 2024-12-14 #673 @maurolepore
Dear @ryanzomorrodi, thanks so much for your submission. I'm happy to be the handling editor.

Within the next few days I'll be reading this thread in more detail, and going through editor checks.

In the meantime, could you suggest three reviewers? It might help you to see [the criteria we use to choose reviewers](https://devdevguide.netlify.app/softwarereview_editor.html#criteria-for-choosing-a-reviewer), particularly [how we define a conflict of interest](https://devdevguide.netlify.app/softwarereview_policies#coi).

Very excited your tool as a way to bridge SAS and R :100:
