### 2024-12-16 7985fec
adding checks
