### 2024-12-16 #673 @ryanzomorrodi
Hi @maurolepore

That's great to hear! I reviewed the criteria, but I'm wondering if there is a set list of reviewers to choose from or should I post something on social media. It looks like the book mentions an Airtable directory, but it feels like that is just for editors.

My primary interest is to select reviewers with some level of SAS experience.
