### 2024-12-18 #673 @maurolepore
@ryanzomorrodi, that's  right, that airtable database is restricted to editors.

If nobody from your own network comes to mind, then no need to stretch to social media.

We ask for suggestions only to ensure we understand the kind of skills the author believes to be useful, so thanks for highlighting the usefulness of SAS experience  🙏👍
