### 2024-12-18 #673 @ryanzomorrodi
Sounds good! I apologize, I'm a relative newcomer to the R open-source world, so I'm not super familiar with the space
