### 2024-12-19 e8e2b36
text lst output for non html format
