### 2024-12-19 1fffd66
Support capture arg for non-html knitr output. #3
