### 2024-12-23 #673 @maurolepore
@ryanzomorrodi thanks again for this submission. I had the change to dive deep into sasquatch and made comments that intend to help it go through review as smoothly as possible.

For that it's crucial that reviewers can explore the package as quickly as possible so that they can invest time in the more interesting aspects of sasquatch (rather than, say, debugging installation issues). Also it will help to make an effort to predict what reviewers will expect, based on what's common in the relevant community of practice. Here I can only contribute what relates to general R-package development, and you'll need to work on the more "SAS" related content.

I acknoledge that the comments are many and can feel intimidating. Please know we're all here to help and we're not in a rush. Take your time to process the comments, ask as many questions as you need, make your own research and judgements. With the holidays ahead I expect the process to experience some delays.

No problem if you already know you won't be able to work for a few months. In that case just let me know and we can tag the issue "on hold" so that in the meantime I can take other packages.

----

---

To help you track my comments I'll tag them with "ml" and a numbered sequence,
e.g. ml01, ml02, and so on. Comments following bullets are for you to consider
-- you may or may not respond to them. Comments following check-boxes are
requests for some action -- please respond.

I already used ml01 above so the next tag will be ml02.

##### **Documentation**

###### Is the case for the package well made?

* ml02. Consider strenghening the case for this package. This is particularly important considering the overlap with other packages (see below). The intent is to articulate why a specific group of people, in a specific situation would find
sasquatch really useful. A useful tool might the a user story. Here's one example based on my relatively shallow understanding (adapt it as necessary):

> As someone who uses R and Quarto but lacks a SAS executable I would like to
have sasquatch so I can try SAS for free in a familiar environment.

This narrower focus in not a problem for rOpenSc; Specific applications in
science is generally what we're looking for.

###### Is the reference index page clear (grouped by topic if necessary)?

Thanks! The index is clearly organized by topic, and with a little more effort
in each help file it will be a pleasure to read:

- [ ] ml03. Please expand the documentation to include more details, particularly
examples. I'll give you one feedback for only one function here but the same
applies to most of them. Consider `install_saspy()`:

* Add a link to the SASPy package.
* Update the website so that the signature in reflects the source code. The
website shows `exatra_packages = NULL` but the source code shows
`extra_packages` has no default.
* Add the type of each argument, e.g.
```
#' @param extra_packages Character. A vector giving the name of one or more
#'   packages to install,
```
* Consider returning something other than `NULL`. See https://design.tidyverse.org/out-invisible.html#whats-the-pattern.
> If a function is called primarily for its side-effects, it should invisibly
return a useful output. If there’s no obvious output, return the first
argument. This makes it possible to use the function with in a pipeline.
* Add examples using the arguments. When are they useful? Can you show
some output? If not, a) show output in a vignette, or b) describe the usage
of each argument in writing, and point to external documentation to learn more
(e.g. SASpy).

###### Are vignettes readable, sufficiently detailed and not just perfunctory?

- [ ] ml04. Please add a vignette to show a more complete usage of the package.
This is particularly useful if the examples don't show output or don't cover
all arguments. Consider a "Get started" vignette. You can create it with `usethis::use_vignette("sasquatch")` and pkgdown should automagically call add a "Get started" hiyper text to the navbar of your pkgdown website. Some of its content might come from what now lives in RADME, and it may point to the other "Setting up" vignette.

##### **Fit**: The package meets criteria for [fit](https://devguide.ropensci.org/policies.html#package-categories) and [overlap](https://devguide.ropensci.org/policies.html#overlap).

- [ ] ml05. Please update the documentation to incorporate the points you brought up in [this comment](https://github.com/ropensci/software-review/issues/673#issuecomment-2495627349). The section [Similar packages](https://ryanzomorrodi.github.io/sasquatch/#similar-packages) lacks a mention all the options you discusses in your comment, and maybe needs to go up -- closer to the goal of the package. A clear way to present this information might be similar to this:

* If you want "a" and have "b", then you may use the native `sas` engine.
* If you want "c" and have "d", then you may use the package 1.
* If you want "e" and have "f", then you may use the package 2.
* If you want "g" and have "h", then you may use sasquatch.

To do this well you'll need to research the perspective of someone with resources
you lack. Where can you find them?

Here are some ideas:

1. Remember how you got involved in SAS. Where was that? Who was there?
1. [rOpenSci's discussion forum](https://discuss.ropensci.org/). SAS is popular
in Pharma and rOpenSci has several folks who work in Pharma. Maybe you can serch
on LinkedIn for people following rOpenSci who work in Pharma?
1. The Pharmaverse: https://pharmaverse.org/. I recently posted in their
[blog](https://pharmaverse.github.io/blog/) and I noticed they are extreemely
friendly and welcoming.

##### **Installation instructions:** Are installation instructions clear enough for human users?

- [ ] ml06. Please refine the installation instructions after watching someone else
walk through the current version. A real user experience can reveal some
unexpected ways in which people interpret the same set of instructions. The
intent is to use effectively the limited time reviewers have. Time they save in
installation they can invest in more interesting aspects of the package.

* ml07. Consider that different people with find easy or hard different things.
Showing many options can be impractical and hard to maintain, but consider
pointing to external resources for users of the three main platforms: Windows,
MacOS and Linux. For me it was smooth to install the system requirements on
Linux from the terminal, but I failed to install via sasquatch:

``` r
packageVersion("sasquatch"); sasquatch::install_saspy()
#> [1] '0.0.0.9017'
#> Error in sasquatch::install_saspy(): argument "extra_packages" is missing, with no default
```

* ml08. Consider mentioning that [pak installs system dependencies automatically](https://pak.r-lib.org/reference/sysreqs.html).
* ml09. Consider recording a video to support installation and setup.

##### **Tests**: If the package has some interactivity / HTTP / plot production etc. are the tests using [state-of-the-art tooling](https://devguide.ropensci.org/building.html#testing)?

I see good effort put into tests, and with some guidance I hope they will be an
excellent way to not only ensure sasquatch does what you expect but also they
will facilitate contributions, including the upcoming reviews.

* ml10. Consider structuring R/ and tests/ such that each exported function f has
an R/f.R and a tests/testthat/test-f.R. This structure became common so it will
likely help reviewers understant each unit of behaviour in relative isolation
from the rest of the system. A good reference is https://r-pkgs.org/testing-basics.html, and I also like https://mastering-shiny.org/scaling-testing.html#testing-functions.

- [ ] ml11. Please ensure to not write files to the developer's system other
than to the standard temporary directory, e.g. via `tempdir()`. This is not only
good etiquette but also it makes cleanup easier. Sometimes a more advanced way
to cleanup becomes necessary, and you may need `on.exit()` or friends from the
[withr package](https://withr.r-lib.org/). See withr in action at
https://testthat.r-lib.org/articles/test-fixtures.html.

- [ ] ml12. Ensure unit tests run quickly (e.g. all tests in 1-5 seconds). This
is critical to allow reviewers to explore and understand sasquatch in the short
time they have. Exploration requires iteration -- lots of it. Also, the faster
your tests are, the more often you'll run them, the faster and more
bugs you'll catch. I acknowledge for sasquatch this is hard because of the
dependency on a remote server. You may need to see how other packages do it,
and dive into advanced testing techniques. Some resources that come to mind are
these:

  * https://r-pkgs.org/testing-advanced.html
  * https://books.ropensci.org/http-testing/
  * https://testthat.r-lib.org/reference/index.html#mocking. You

- [ ] ml13. Whenever possible please replace printed output with something that's
easier to handle (e.g. `message()` can be suppressed with
`suppressedMessages()`). The clutter in the test output will make it hard for
reviewers to see what really matters. See
https://ropensci.org/blog/2024/02/06/verbosity-control-packages/.

<details><summary>Example of the test output I see</summary>

```
==> devtools::test()

ℹ Testing sasquatch
✔ | F W  S  OK | Context
⠏ |          0 | connection   SAS Connection established.
⠋ |          1 | connection   SAS Connection established.
⠙ |          2 | connection   SAS Connection terminated.
✖ | 1        5 | connection [29.6s]
──────────────────────────────
Failure (test-connection.R:16:3): test connection
`sas_get_session()` did not throw the expected error.
──────────────────────────────
⠏ |          0 | conversion   SAS Connection established.
⠋ | 1        0 | conversion   SAS Connection established.
✖ | 2 1      0 | conversion [35.3s]
──────────────────────────────
Error (test-conversion.R:9:3): SAS to R data.frame
Error in `sas_to_r("iris", "sashelp")`: object 'x' not found
    ▆
 1. └─sasquatch::sas_to_r("iris", "sashelp") at test-conversion.R:9:3
 2.   └─reticulate::py_to_r(x) at sasquatch/R/conversion.R:61:3
 3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
 4.     └─reticulate:::py_to_r_cpp(x)

Warning (test-conversion.R:36:3): R to SAS data.frame
`x` rownames will not be transferred as columns.
    ▆
 1. └─sasquatch::r_to_sas(iris, "iris") at test-conversion.R:36:3
 2.   └─sasquatch:::chk_has_rownames(x) at sasquatch/R/conversion.R:22:3
 3.     └─chk::wrn(x_name, " rownames will not be transferred as columns") at sasquatch/R/additional_chk.R:14:3

Error (test-conversion.R:37:3): R to SAS data.frame
Error in `sas_to_r("iris")`: object 'x' not found
    ▆
 1. └─sasquatch::sas_to_r("iris") at test-conversion.R:37:3
 2.   └─reticulate::py_to_r(x) at sasquatch/R/conversion.R:61:3
 3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
 4.     └─reticulate:::py_to_r_cpp(x)
──────────────────────────────
⠙ | 1        1 | io           SAS Connection established.
⠹ | 1        2 | io           SAS Connection terminated.
⠼ | 1        4 | io           SAS Connection established.
SAS Connection terminated.
⠴ | 1        5 | io           SAS Connection established.
⠧ | 1        7 | io           SAS Connection terminated.
⠇ | 1        8 | io           SAS Connection established.
SAS Connection terminated.
⠏ | 1        9 | io           SAS Connection established.
⠙ | 1       11 | io           SAS Connection terminated.
⠹ | 1       12 | io           SAS Connection established.
⠸ | 1       13 | io           SAS Connection terminated.
✖ | 1       13 | io [227.9s]
──────────────────────────────
Failure (test-io.R:17:3): uploading files to SAS
`sas_upload(local_path, sas_path)` did not throw the expected error.
──────────────────────────────
⠋ |          1 | run          SAS Connection established.
⠙ |          2 | run          SAS Connection terminated.
⠹ |          3 | run          SAS Connection established.
⠦ |          7 | run          SAS Connection terminated.
✔ |          7 | run [62.1s]

══ Results ═══════════════════

── Failed tests ──────────────
Failure (test-connection.R:16:3): test connection
`sas_get_session()` did not throw the expected error.

Error (test-conversion.R:9:3): SAS to R data.frame
Error in `sas_to_r("iris", "sashelp")`: object 'x' not found
    ▆
 1. └─sasquatch::sas_to_r("iris", "sashelp") at test-conversion.R:9:3
 2.   └─reticulate::py_to_r(x) at sasquatch/R/conversion.R:61:3
 3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
 4.     └─reticulate:::py_to_r_cpp(x)

Error (test-conversion.R:37:3): R to SAS data.frame
Error in `sas_to_r("iris")`: object 'x' not found
    ▆
 1. └─sasquatch::sas_to_r("iris") at test-conversion.R:37:3
 2.   └─reticulate::py_to_r(x) at sasquatch/R/conversion.R:61:3
 3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
 4.     └─reticulate:::py_to_r_cpp(x)

Failure (test-io.R:17:3): uploading files to SAS
`sas_upload(local_path, sas_path)` did not throw the expected error.

```

##### **Contributing information**: Is the documentation for contribution clear enough e.g. tokens for tests, playgrounds?

- [ ] ml14. Please add instructions for reviewers and other contributors to
successfullly run tests. What seems like test setup points to a specific,
private `sas_user` that will unlikely work for reviewers. See https://devguide.ropensci.org/maintenance_collaboration.html. Also note that
testthat enforces a different name-convention for heper files versus setup files.
See https://testthat.r-lib.org/articles/special-files.html#setup-files.

##### Additional considerations

- [ ] ml15. Please ensure all URLs are correct. You may use `urlchecker::url_check()`.

``` r
urlchecker::url_check()
#> fetching [  0 / 15 ]fetching [  1 / 15 ]fetching [  2 / 15 ]fetching [  3 / 15 ]fetching [  4 / 15 ]fetching [  5 / 15 ]fetching [  6 / 15 ]fetching [  7 / 15 ]fetching [  8 / 15 ]fetching [  9 / 15 ]fetching [ 10 / 15 ]fetching [ 11 / 15 ]fetching [ 12 / 15 ]fetching [ 13 / 15 ]fetching [ 14 / 15 ]
#> ! Warning: DESCRIPTION:26:50 Moved
#> URL: https://ryanzomorrodi.github.io/sasquatch/, https://github.com/ryanzomorrodi/sasr
#>                                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#>                                                  https://github.com/ryanzomorrodi/sasquatch
#> ! Warning: DESCRIPTION:33:13 Moved
#> BugReports: https://github.com/ryanzomorrodi/sasr/issues
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#>             https://github.com/ryanzomorrodi/sasquatch
#> ! Warning: README.md:8:17 Moved
#> [![R-CMD-check](https://github.com/ryanzomorrodi/sasr/actions/workflows/R-CMD-check.yaml/badge.svg)](https://github.com/ryanzomorrodi/sasr/actions/workflows/R-CMD-check.yaml)
#>                 ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#>                 https://github.com/ryanzomorrodi/sasquatch/actions/workflows/R-CMD-check.yaml
#> ! Warning: DESCRIPTION:33:13 Moved
#> BugReports: https://github.com/ryanzomorrodi/sasr/issues
#>             ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#>             https://github.com/ryanzomorrodi/sasquatch/issues
#> Warning in readLines(file_path): incomplete final line found on
#> '/home/rstudio/git/sasquatch/vignettes/setting_up.rmd'
#> ✖ Error: vignettes/setting_up.rmd:17:46 404: Not Found
#> `sasquatch` works by utilizing the [`SASPy`](https://sassoftware.github.io/SASPy/) python package, similar to packages like [`sasr`](https://github.com/insightsengineering/sasr) or [configSAS](https://github.com/baselr/configSAS). This means everything we do to connect R and SAS, needs to go through `SASPy`.
#>                                              ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> ✖ Error: vignettes/setting_up.rmd:21:120 404: Not Found
#> For more information about setting up non-ODA `SASPy` connections, check out the [`SASPy` configuration documentation](https://sassoftware.github.io/SASPy/configuration.html#).
#>                                                                                                                        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> Warning in readLines(file_path): incomplete final line found on
#> '/home/rstudio/git/sasquatch/vignettes/setting_up.rmd'
#> ✖ Error: vignettes/setting_up.rmd:21:120 404: Not Found
#> For more information about setting up non-ODA `SASPy` connections, check out the [`SASPy` configuration documentation](https://sassoftware.github.io/SASPy/configuration.html#).
#>                                                                                                                        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> Warning in readLines(file_path): incomplete final line found on
#> '/home/rstudio/git/sasquatch/vignettes/setting_up.rmd'
#> ✖ Error: vignettes/setting_up.rmd:52:88 403: Forbidden
#> Make sure Java is installed on your system. You can download Java from [their website](https://www.java.com/en/download/). Note the Java installation path.
#>                                                                                        ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> Warning in readLines(file_path): incomplete final line found on
#> '/home/rstudio/git/sasquatch/vignettes/setting_up.rmd'
#> ✖ Error: vignettes/setting_up.rmd:40:98 404: Not Found
#> Make sure Python is installed on your system. You can download from [Python Software Foundation](https://www.python.org/downloads'), or use `install_python()` from reticulate:
#>                                                                                                  ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#> ! Warning: README.md:29:12 Moved
#> Academics](https://www.sas.com/en_us/software/on-demand-for-academics)
#>            ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
#>            https://www.sas.com/content/sascom/en_us/software/on-demand-for-academics/
```

* ml16. Consider showing an example of the round trip from `r_to_sas()` and
`sas_to_r()`. If the round trip is complete, the final result should be
identical to the starting point.

* ml17. Consider adding the `sas_` prefix to the few functions that lack it, e.g.
`r_to_sas()` may become `sas_from_r()`.

* ml18. Consider running a global search for RStudio and making sure wording won't
be obsolete when Positron becomes more popular.
