### 2024-12-23 bea6b7a
Fix rownames check
