+++ R/additional_chk.R
@@ -11,6 +11,6 @@ chk_has_rownames <- function(x, x_name = NULL) {
     return(invisible(x))
   }
   if (is.null(x_name)) x_name <- chk::deparse_backtick_chk(substitute(x))
-  chk::wrn(x_name, " rownames will not be transferred as columns")
+  chk::wrn(x_name, " rownames will not be transferred as a column")
 }
-vld_has_rownames <- function(x) is.na(.row_names_info(x)[1])
+vld_has_rownames <- function(x) is.na(.row_names_info(x, type = 0)[1])
