### 2024-12-23 5132cc7
Make connection print out messages
