### 2024-12-23 faace65
Bug: allow install of no additional packages
