### 2024-12-23 e739051
Document connection checks
