+++ R/run.R
@@ -17,7 +17,7 @@
 sas_run_string <- function(input) {
   chk_connection()
 
-  execute_safely(
+  execute_if_connection_active(
     results <- .pkgenv$session$submit(input)
   )
 
