+++ R/utils.R
@@ -45,17 +45,3 @@ write_file <- function(output, path, overwrite) {
 
   cat(output, file = path)
 }
-
-#' Execute saspy function with a connection check if the function fails
-execute_safely <- function(code) {
-  calling_env <- parent.frame()
-  tryCatch({
-    code
-  }, error = function(e) {
-    if (is.null(.pkgenv$session$SASpid)) {
-      chk::abort_chk("No active SAS session. Use sas_connect() to start one.", call = calling_env)
-    } else {
-      e
-    }
-  })
-}
