### 2024-12-25 84b6cb6
Add documentation and type checking to conversion  and connection functions
