+++ R/connection.R
@@ -3,8 +3,14 @@
 #' @description
 #' Starts a SAS session. This is required before doing anything!
 #' 
-#' @param config Configuration to use from the SAS_config_names list in 
-#' `sascfg_personal.py`.
+#' @param cfgname string; Name of configuration to use from the SAS_config_names
+#' list in `sascfg_personal.py`.
+#' 
+#' @details
+#' All configurations are specified within the `sascfg_personal.py` file inside
+#' the `SASPy` package. For more information about `SASPy` configuration, check
+#' out the [configuration documentation](https://sassoftware.github.io/saspy/configuration.html)
+#' or `vignette("setting_up")`.
 #' 
 #' @return No return value.
 #' 
@@ -12,20 +18,20 @@
 #' 
 #' @examples
 #' \dontrun{
-#' sas_connect(config = "oda")
+#' sas_connect(cfgname = "oda")
 #' }
-sas_connect <- function(config) {
-  if (!missing(config)) {
-    chk::chk_string(config)
+sas_connect <- function(cfgname) {
+  if (!missing(cfgname)) {
+    chk::chk_string(cfgname)
   }
   
-  if (missing(config)) {
+  if (missing(cfgname)) {
     reticulate::py_capture_output(
       .pkgenv$session <- .pkgenv$SASPy$SASsession()
     )
   } else {
     reticulate::py_capture_output(
-      .pkgenv$session <- .pkgenv$SASPy$SASsession(config = config)
+      .pkgenv$session <- .pkgenv$SASPy$SASsession(cfgname = cfgname)
     )
   }
   chk::msg("SAS Connection established.")
@@ -60,10 +66,21 @@ sas_disconnect <- function() {
 
 #' Get current SAS session
 #' 
-#' Gets current SAS session so that you can use functions not yet implemented.
-#' Can also be useful for testing or using Python.
+#' Returns the current SAS session, which can be used to extend `sasquatch`
+#' functionality or access the current session within Python.
 #' 
 #' @return Current SAS session.
+#' 
+#' @details
+#' ## Extending `sasquatch` functionality
+#' `SASPy` has a wealth of functionality that has not been implemented within 
+#' `sasquatch`. `sas_get_session()` offers a gateway to unimplemented
+#' functionality within the [SASsession class](https://sassoftware.github.io/saspy/api.html#sas-session-object).
+#' 
+#' ## Using Python
+#' When utilizing Python, R, and SAS, start the session within R using 
+#' `sas_connect()` and utilize `reticulate` to pass the current SAS session
+#' to Python.
 #' 
 #' @export
 #' 
@@ -84,6 +101,8 @@ sas_get_session <- function() {
 #' @details 
 #' Use `execute_if_connection_active()` for any function that relies on a SAS
 #' connection to catch inactive sessions.
+#' 
+#' @keywords internal
 chk_connection <- function() {
   if (vld_connection()) {
     return(invisible())
@@ -101,6 +120,8 @@ vld_connection <- function() exists("session", envir = .pkgenv) && !is.null(.pkg
 #' The SAS connection is asynchronous so it can become inactive without the user
 #' knowing. When the connection is inactive and an action is preformed it will 
 #' check the connection and raise an error.
+#' 
+#' @keywords internal
 execute_if_connection_active <- function(code) {
   calling_env <- parent.frame()
   tryCatch({
