### 2024-12-26 9900dfd
Add file management documentation and fix return values
