+++ man/sas_list.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/io.R
+% Please edit documentation in R/file.R
 \name{sas_list}
 \alias{sas_list}
 \title{List contents of a SAS directory}
@@ -10,7 +10,7 @@ sas_list(path)
 \item{path}{Path of directory on remote SAS server to list the contents of.}
 }
 \value{
-No return value.
+Vector of file or directory names.
 }
 \description{
 Lists the files or directories of a directory within the remote SAS server.
