+++ man/sas_upload.Rd
@@ -1,27 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/io.R
-\name{sas_upload}
-\alias{sas_upload}
-\title{Upload a file to SAS}
-\usage{
-sas_upload(local_path, sas_path)
-}
-\arguments{
-\item{local_path}{Path of file on local machine to be uploaded.}
-
-\item{sas_path}{Path to upload local file to on the remote SAS server.}
-}
-\value{
-No return value.
-}
-\description{
-Uploads a file to the remote SAS server.
-}
-\examples{
-\dontrun{
-sas_connect()
-
-quarto_file <- system.file("sasquatch.qmd", package = "sasquatch")
-sas_upload(quarto_file, "sasquatch.qmd")
-}
-}
