### 2024-12-26 bd69487
Change setting up documentation to more broad configuration
