### 2024-12-26 05f1267
Add more package comparison to README
