### 2024-12-26 78ecacd
Add vignette covering basic functionality
