+++ vignettes/sasquatch.Rmd
@@ -1,0 +1,230 @@
+---
+title: "Getting Started"
+output: rmarkdown::html_vignette
+vignette: >
+  %\VignetteIndexEntry{sasquatch}
+  %\VignetteEngine{knitr::rmarkdown}
+  %\VignetteEncoding{UTF-8}
+---
+
+```{r, include = FALSE}
+knitr::opts_chunk$set(
+  collapse = TRUE,
+  comment = "#>"
+)
+```
+
+```{r, eval = FALSE}
+library(sasquatch)
+```
+
+## Session management
+
+Every `sasquatch` script starts with `sas_connect()`. By default, you will be
+connected to your first SAS configuration (for more about configurations see
+`vignette("configuration")`).
+
+```{r, eval = FALSE}
+sas_connect()
+#> SAS Connection established.
+```
+
+You can also specify your configuration by name:
+
+```{r, eval = FALSE}
+sas_connect(cfgname = "my_config")
+#> SAS Connection established.
+```
+
+If you ever need to end a connection, you can do so by using:
+
+```{r, eval = FALSE}
+sas_disconnect()
+#> SAS Connection terminated.
+```
+
+All connection information is stored within a `saspy.sasbase.SASsession` object.
+For most individuals, you will never need to interact with this object, but if
+you would like to implement functionality not currently within `sasquatch` or 
+would like to access the current session from Python, you can via:
+
+```{r, eval = FALSE}
+sas_get_session()
+#> Access Method         = IOM
+#> SAS Config name       = my_config
+#> SAS Config file       = /home/user/.virtualenvs/r-saspy/lib/python3.12/site-packages/saspy/sascfg_personal.py
+#> WORK Path             = /saswork/SAS_work0D3600010B4E_odaws01-usw2-2.oda.sas.com/SAS_work920700010B4E_odaws01-usw2-2.oda.sas.com/
+#> SAS Version           = 9.04.01M7P08062020
+#> SASPy Version         = 5.101.1
+#> Teach me SAS          = False
+#> Batch                 = False
+#> Results               = Pandas
+#> SAS Session Encoding  = utf-8
+#> Python Encoding value = utf-8
+#> SAS process Pid value = 68430
+#> SASsession started    = Wed Dec 25 13:58:22 2024
+```
+
+## Executing SAS code
+
+You can execute SAS code in a variety of different ways. All `sas_run_*()` 
+functions generate htmlwidgets, which display both the output and log.
+
+To execute a string of SAS code use `sas_run_string()`
+
+```{r, eval = FALSE}
+sas_run_string("PROC MEANS DATA = sashelp.cars;RUN;")
+```
+
+```{r, echo = FALSE}
+readRDS("files/proc_means_widget.rds")
+```
+
+To execute a SAS script use `sas_run_file()`
+
+```{r, eval = FALSE}
+cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")
+sas_widget <- sas_run_file("script.sas")
+```
+
+### Within quarto
+
+Quarto documents are a great way to use SAS and R together because they couple
+R and SAS within a single reproducable document.
+
+You can create a `sasquatch` quarto document by:
+
+* Specifying the format as html (currenly, only html is well supported)
+* Specifying the engine as knitr
+* Creating an R code block with `library(sasquatch); sas_connect()`
+
+Now, SAS code can be contained within "sas" code blocks.
+
+`````{verbatim}
+---
+format: html
+engine: knitr
+---
+
+```{r}
+library(sasquatch)
+sas_connect()
+```
+
+```{sas}
+PROC MEANS DATA = sashelp.cars;
+RUN;
+```
+`````
+
+#### RStudio
+
[…]
