### 2024-12-26 adc900a
Show r <-> sas round trip in getting started vignette
