### 2024-12-28 b16fd07
non-html markdownify knitr implementation
