+++ _pkgdown.yml
@@ -19,11 +19,12 @@ reference:
   - sas_disconnect
   - sas_get_session
 
-- title: Executing Code
+- title: Code Execution
   contents:
   - sas_run_string
   - sas_run_selected
   - sas_run_file
+  - sas_set_tempdir
   - sas_engine
 
 - title: R <-> SAS
