+++ R/file.R
@@ -62,9 +62,7 @@ sas_file_upload <- function(local_path, sas_path) {
 #' }
 sas_file_download <- function(sas_path, local_path) {
   chk_connection()
-  chk::chk_not_missing(sas_path)
   chk::chk_string(sas_path)
-  chk::chk_not_missing(local_path)
   chk::chk_string(local_path)
 
   execute_if_connection_active(
@@ -142,7 +140,8 @@ sas_file_remove <- function(path) {
 #' }
 sas_file_copy <- function(from_path, to_path) {
   chk_connection()
-  chk::chk_string(path)
+  chk::chk_string(from_path)
+  chk::chk_string(to_path)
 
   execute_if_connection_active(
     result <- .pkgenv$session$file_copy(from_path, to_path)
@@ -153,6 +152,43 @@ sas_file_copy <- function(from_path, to_path) {
   }
 
   result$Success
+}
+
+#' Check if file on SAS exists
+#' 
+#' Checks if a file exists on the remote SAS server. Is analogous to 
+#' `file.exists()`, but for the remote SAS server.
+#' 
+#' @param path Path of file on remote SAS server.
+#' 
+#' @return Logical value indicating if file exists on remote SAS server.
+#' 
+#' @export
+#' 
+#' @examples
+#' \dontrun{
+#' # connect to SAS
+#' sas_connect()
+#' 
+#' # create a file and upload it to SAS
+#' cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")
+#' sas_file_upload(local_path = "script.sas", sas_path = "~/script.sas")
+#' 
+#' # check if file exists on SAS
+#' sas_file_exists("~/script.sas")
+#' }
+sas_file_exists <- function(path) {
+  chk_connection()
+  chk::chk_string(path)
+
+  file <- basename(path)
+  dir <- gsub(file, "", path)
+
+  execute_if_connection_active(
+    dirlist <- .pkgenv$session$dirlist(dir)
+  )
+
+  any(dirlist == file)
 }
 
 #' List contents of a SAS directory
@@ -177,5 +213,5 @@ sas_list <- function(path) {
   execute_if_connection_active(
     dirlist <- .pkgenv$session$dirlist(path)
   )
-  dirlist
+  as.vector(dirlist)
 }
