+++ man/sas_file_exists.Rd
@@ -1,0 +1,31 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/file.R
+\name{sas_file_exists}
+\alias{sas_file_exists}
+\title{Check if file on SAS exists}
+\usage{
+sas_file_exists(path)
+}
+\arguments{
+\item{path}{Path of file on remote SAS server.}
+}
+\value{
+Logical value indicating if file exists on remote SAS server.
+}
+\description{
+Checks if a file exists on the remote SAS server. Is analogous to
+\code{file.exists()}, but for the remote SAS server.
+}
+\examples{
+\dontrun{
+# connect to SAS
+sas_connect()
+
+# create a file and upload it to SAS
+cat("PROC MEANS DATA = sashelp.cars;RUN;", file = "script.sas")
+sas_file_upload(local_path = "script.sas", sas_path = "~/script.sas")
+
+# check if file exists on SAS
+sas_file_exists("~/script.sas")
+}
+}
