+++ NAMESPACE
@@ -8,6 +8,7 @@ export(sas_disconnect)
 export(sas_engine)
 export(sas_file_copy)
 export(sas_file_download)
+export(sas_file_exists)
 export(sas_file_remove)
 export(sas_file_upload)
 export(sas_get_session)
