### 2024-12-28 df65c04
Initial revamping of test suit
