+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   configuration: configuration.html
   sasquatch: sasquatch.html
-last_built: 2024-12-28T01:16Z
+last_built: 2024-12-28T20:25Z
 urls:
   reference: https://ryanzomorrodi.github.io/sasquatch/reference
   article: https://ryanzomorrodi.github.io/sasquatch/articles
