+++ tests/testthat/test-conversion.R
@@ -15,15 +15,15 @@ test_that("Back and forth", {
   )
 
   "list columns: R to SAS"
-  expect_error(r_to_sas(df_all, "df"), "must only have logical, integer, double, factor, character, POSIXct, or Date class columns")
+  expect_error(sas_from_r(df_all, "df"), "must only have logical, integer, double, factor, character, POSIXct, or Date class columns")
 
   "valid columns: R to SAS"
   df <- df_all[!sapply(df_all, is.list)]
-  expect_equal(r_to_sas(df, "df"), df)
+  expect_equal(sas_from_r(df, "df"), df)
 
   "rownames: R to SAS"
   rownames(df) <- paste("row", 1:3)
-  expect_warning(r_to_sas(df, "df"), "rownames will not be transferred as a column")
+  expect_warning(sas_from_r(df, "df"), "rownames will not be transferred as a column")
   rownames(df) <- NULL
 
   "back to R check"
