+++ tests/testthat/test-from-r.R
@@ -1,6 +1,7 @@
-test_that("back and forth", {
+test_that("R data.frame checks", {
   skip_on_cran()
   skip_if_offline()
+  withr::defer(sas_get_session()$submit("proc datasets library=WORK;delete df;run;"))
 
   df_all <- data.frame(
     a = c(1, 2.5, NA),
@@ -14,17 +15,33 @@ test_that("back and forth", {
     i = I(list(list(1, 2:3), list(4:6), list(NA)))
   )
 
-  "list columns: R to SAS"
+  "list columns error"
   expect_error(sas_from_r(df_all, "df"), "must only have logical, integer, double, factor, character, POSIXct, or Date class columns.", fixed = TRUE)
 
-  "valid columns: R to SAS"
+  "rownames warning"
   df <- df_all[!sapply(df_all, is.list)]
-  expect_equal(sas_from_r(df, "df"), df)
-
-  "rownames: R to SAS"
   rownames(df) <- paste("row", 1:3)
   expect_warning(sas_from_r(df, "df"), "rownames will not be transferred as a column", fixed = TRUE)
-  rownames(df) <- NULL
+})
+
+test_that("Round trip", {
+  skip_on_cran()
+  skip_if_offline()
+  withr::defer(sas_get_session()$submit("proc datasets library=WORK;delete df;run;"))
+
+  df <- data.frame(
+    a = c(1, 2.5, NA),
+    b = c(1:2, NA),
+    c = c(T, F, NA),
+    d = c("a", "b", NA),
+    e = factor(c("a", "b", NA)),
+    f = as.Date("2015-12-09") + c(1:2, NA),
+    g = as.POSIXct("2015-12-09 10:51:34.5678", tz = "UTC") + c(1:2, NA)
+  )
+
+  "valid columns"
+  expect_equal(sas_from_r(df, "df"), df)
+  expect_true(sas_get_session()$exist("df", libref = "WORK"))
 
   "back to R check"
   df_from_sas <- sas_to_r("df")
