+++ tests/testthat/test-to-r.R
@@ -1,0 +1,75 @@
+test_that("SAS to R data.frame", {
+  skip_on_cran()
+  skip_if_offline()
+
+  r_airline <- structure(list(
+    DATE = c("1949-01-01", "1949-02-01", "1949-03-01", 
+      "1949-04-01", "1949-05-01", "1949-06-01", "1949-07-01", "1949-08-01", 
+      "1949-09-01", "1949-10-01", "1949-11-01", "1949-12-01", "1950-01-01", 
+      "1950-02-01", "1950-03-01", "1950-04-01", "1950-05-01", "1950-06-01", 
+      "1950-07-01", "1950-08-01", "1950-09-01", "1950-10-01", "1950-11-01", 
+      "1950-12-01", "1951-01-01", "1951-02-01", "1951-03-01", "1951-04-01", 
+      "1951-05-01", "1951-06-01", "1951-07-01", "1951-08-01", "1951-09-01", 
+      "1951-10-01", "1951-11-01", "1951-12-01", "1952-01-01", "1952-02-01", 
+      "1952-03-01", "1952-04-01", "1952-05-01", "1952-06-01", "1952-07-01", 
+      "1952-08-01", "1952-09-01", "1952-10-01", "1952-11-01", "1952-12-01", 
+      "1953-01-01", "1953-02-01", "1953-03-01", "1953-04-01", "1953-05-01", 
+      "1953-06-01", "1953-07-01", "1953-08-01", "1953-09-01", "1953-10-01", 
+      "1953-11-01", "1953-12-01", "1954-01-01", "1954-02-01", "1954-03-01", 
+      "1954-04-01", "1954-05-01", "1954-06-01", "1954-07-01", "1954-08-01", 
+      "1954-09-01", "1954-10-01", "1954-11-01", "1954-12-01", "1955-01-01", 
+      "1955-02-01", "1955-03-01", "1955-04-01", "1955-05-01", "1955-06-01", 
+      "1955-07-01", "1955-08-01", "1955-09-01", "1955-10-01", "1955-11-01", 
+      "1955-12-01", "1956-01-01", "1956-02-01", "1956-03-01", "1956-04-01", 
+      "1956-05-01", "1956-06-01", "1956-07-01", "1956-08-01", "1956-09-01", 
+      "1956-10-01", "1956-11-01", "1956-12-01", "1957-01-01", "1957-02-01", 
+      "1957-03-01", "1957-04-01", "1957-05-01", "1957-06-01", "1957-07-01", 
+      "1957-08-01", "1957-09-01", "1957-10-01", "1957-11-01", "1957-12-01", 
+      "1958-01-01", "1958-02-01", "1958-03-01", "1958-04-01", "1958-05-01", 
+      "1958-06-01", "1958-07-01", "1958-08-01", "1958-09-01", "1958-10-01", 
+      "1958-11-01", "1958-12-01", "1959-01-01", "1959-02-01", "1959-03-01", 
+      "1959-04-01", "1959-05-01", "1959-06-01", "1959-07-01", "1959-08-01", 
+      "1959-09-01", "1959-10-01", "1959-11-01", "1959-12-01", "1960-01-01", 
+      "1960-02-01", "1960-03-01", "1960-04-01", "1960-05-01", "1960-06-01", 
+      "1960-07-01", "1960-08-01", "1960-09-01", "1960-10-01", "1960-11-01", 
+      "1960-12-01"), 
[…]
