+++ tests/testthat/helper-setup.R
@@ -1,3 +1,0 @@
-sas_username <- "u63378825"
-test_number <- ceiling(runif(1, 0, 10000000)) # helps make sure CI tests don't interfere with one another
-sas_connect()
