### 2024-12-29 39d2b32
Add types to documentation
