+++ R/connect.R
@@ -4,13 +4,13 @@
 #' Starts a SAS session. This is required before doing anything!
 #' 
 #' @param cfgname string; Name of configuration to use from the SAS_config_names
-#' list in `sascfg_personal.py`.
+#' list within in `sascfg_personal.py`.
 #' 
 #' @details
 #' All configurations are specified within the `sascfg_personal.py` file inside
 #' the `SASPy` package. For more information about `SASPy` configuration, check
 #' out the [configuration documentation](https://sassoftware.github.io/saspy/configuration.html)
-#' or `vignette("setting_up")`.
+#' or `vignette("configuration")`.
 #' 
 #' @return No return value.
 #' 
