+++ R/to_r.R
@@ -30,6 +30,7 @@
 #' }
 sas_to_r <- function(table_name, libref = "WORK")  {
   chk_connection()
+  chk::chk_not_missing(table_name, "`table_name`")
   chk::chk_string(table_name)
   chk::chk_string(libref)
   
