### 2024-12-30 #673 @ryanzomorrodi
Hi @maurolepore

Sorry for the delay in response. I have tried to address as many of the issues that I can. I will probably need some time to reach out to others to see if there is anyone within the pharmaverse or rOpenSci interested in telling me about their use cases.

**ml02**

I try to kind of get at the purpose with the following the text:

> sasquatch allows you to combine the power of R, SAS, and quarto together to create reproducible multilingual reports. sasquatch can:
>
> * Run SAS code blocks interactively
> * Send data back and forth between SAS and R
> * Conduct basic file management on a SAS client
> * Render SAS output within quarto documents.

I could expand on it further, but I think the use case for thispackage is summarized within the first line. I see this as useful for someone looking to create multilingual reports where they can connect to nearly any SAS client, pass data back and forth, and have a single reproducable document contain all the code and output.

**ml03**

I added types to every function parameter, specified analagous functions for `sas_file_*` functions, and provided additional detail to functions that required additional info like `sas_engine()`.

**ml04**

I created a [Getting started vignette](https://ryanzomorrodi.github.io/sasquatch/articles/sasquatch.html) and included a brief breakdown of each function.

**ml05**

I did add additional info to [similar packages](https://ryanzomorrodi.github.io/sasquatch/#comparison-with-similar-packages) section of the README.

I really don't see a reason to use `configSAS` or `sasr` over `sasquatch` because `sasquatch` encorporates all the same functionality plus more. `SASmarkdown` on the other hand is easier to set up and does a better job at rendering pdf output, so if you don't care about interactive SAS support or passing data back and forth, I would recommend that. So that's what I tried to articulate.

That being said I think getting outside perspectives is probably a good thing. Hopefully, I can find interested individuals or even get the views of my professors who use SAS.

**ml06**

This I have yet to do, but plan on doing.

**ml07**

This error was a very simple mistake in the function code. I have since implemented a lot more testing for many of the functions, so hopefully I caught most of the mistakes like this. I'm not sure if you have any recommendations as to how I can better test the `install_saspy()` and `configure_saspy()` functions. Because they require changing system files I think it would be very hard to test them and revert the global state.

I also use linux so maybe watching others on other platforms would help me get a better idea of potential pitfalls.

**ml08**

I may add this later. I chose pak for the purpose that it explictly lists out all system dependencies needed, which is definitely and added benefit.

**ml09**

Recording an installation video is definitely something I could do. I first want to see what different platform's installation is like, so I might delay this.

**ml10**

Done, initially I liked having relevant functions within the same file because they were close together, but even when separating them out they still organized themselves.

**ml12**

Right now all the tests run in about 2 minutes. I am missing tests for `install_saspy()` and `configure_saspy()`, but I think two minutes is reasonable. The main rate limiting step is `sas_connect()`, which takes a while to set up. Instead of connecting and disconnecting for each function to see if the connection check works within the function, I just set up tests for the connection check function and didn't bother with setting up checks for each individual function.

**ml13**

There were two uses of `cat()` which I replaced with `message()`.

**ml14**

I added instructions that specifies that testing can be done with SAS ODA and that `install_saspy()` and `configure_saspy(template = "oda")` should be used to set it up.

I removed the setup file entirely.

**ml15**

I used `urlchecker::url_check()` and fixed the issues that popped up, but for some reason https://www.python.org/downloads/ kept giving a 404 within the checker. I left it in though, because I cannot reproduce it within the browser.

**ml16**

A SAS <-> R was added to the Getting Started vignette. There is no way to make the round trip perfect because of limitations with SAS data types, but I explained any differences.

**ml17**

Changed `r_to_sas()` to `sas_from_r()`.

**ml18**

I use Positron, but I have made most of the docs with RStudio in mind because at least for now that is the more dominant among R users, but I tried to include references to Positron, where functionality diverged.

Thank you for all your thorough comments. This is all helpful insight!
