### 2024-12-30 eb08f47
Add run_file checks
