### 2024-12-31 #673 @maurolepore
@ryanzomorrodi thanks so much for your work!

Based on your comment I ticked lots of boxes.  In the next few days I'll try installation, setup, and testing and I'll think about your question:

> I'm not sure if you have any recommendations as to how I can better test the `install_saspy()` and `configure_saspy()` functions. Because they require changing system files I think it would be very hard to test them and revert the global state.

I'll come back early next year :smile:

Happy holidays!

----

BTW, re-reading my own comment above I found so many sloppy typos -- I apologize.
