### 2025-01-13 #673 @maurolepore
@ryanzomorrodi, thanks again for the work you put into enhancing the code and sharing it with rOpenSci. I looked at your answers and as I looked at the code I made a number of notes that you might find helpful. At this point most of them should be thought as optional suggestions.

Again, remember the intent is to ensure the reviewers can do their job as quickly as possible, so please try your best to maximize the chance they'll be able to install, configure and test the code without errors, and as quickly as possible.  I share some ideas to make tests run faster.

## README

* ml19. Keep the top section focused on the goal

The top section reads:

> sasquatch relies on the SASPy Python package and the reticulate R package to interoperate with Python. Check out vignette("configuration") for guidance on SASPy configuration.

Consider reserving this section for the package goal alone, and moving configuration details after installation. These seems to be a natural sequence:

1. Decide I want this package,
2. Install it,
3. Set it  up,
4. Use it.

* ml20. Add configuration for beginners

Before usage we need configuration. Sending readears to a vignette is fine but making README self-contained seems better because it minimizes the information-hunt before a succesfull first run.

Consider moving here the section "SAS On Demand for Academics", explaining this is a good way to get started for those who lack a license (which seems to be an important audience of this package), then pointing to the vignette for more advanced configuration.

I would focus on helping a) the reviewers, b) then curious R users, c) then anyone else. The users a-b likely lack a SAS license and are rushed to get something running.

* ml21. Did you mean quarto?

I see: "Now, you should be able to run SAS code blocks in RStudio like any other."
Consider: "Now, you should be able to run SAS code blocks in Quarto like any other."

Also, sasquatch seems to also work with Rmarkdown, right? Does it deserve a mention?

## Enhancements

* ml22. Consider using `memoise::memoise()` to cache the result of slow function within an R session. (This is suiteble only when you two calls with the same arguments should always return the same result within an R session.) This should save time for reviewers and users.

``` r
devtools::load_all()
#> ℹ Loading sasquatch
sas_connect()
#> SAS Connection established.

system.time(sas_list("bad"))
#>    user  system elapsed
#>   0.003   0.001   2.006
# Same time
system.time(sas_list("bad"))
#>    user  system elapsed
#>   0.004   0.000   2.708

# Maybe this specific funciton isn't the best candidate for memoise() but
# hopefully illustrates the idea
sas_list2 <- memoise::memoise(sas_list)

system.time(sas_list2("bad"))
#>    user  system elapsed
#>   0.005   0.001   3.011
# No time
system.time(sas_list2("bad"))
#>    user  system elapsed
#>   0.013   0.000   0.012
system.time(sas_list2("bad"))
#>    user  system elapsed
#>   0.000   0.000   0.001
system.time(sas_list2("bad"))
#>    user  system elapsed
#>       0       0       0

# https://memoise.r-lib.org/
# https://r-pkgs.org/testing-advanced.html
```

* ml23. Consider modifying `sas_connect()` to avoid unintentional reconnections and saves time. This may make tests and examples run faster, saving reviewer time.

``` r
devtools::load_all()
#> ℹ Loading sasquatch

# Refactor: Rename the implementation
sas_connect_impl <- sas_connect

# Enhance the user-facing function
sas_connect <- function(cfgname, reconnect = FALSE) {
  unconnected <- is.null(sas_get_session())

  if (unconnected || reconnect) {
    sas_connect_impl(cfgname)
  } else {
    message("Already connected.")
    invisible()
  }
}

# The current implementation reconnects, which is slow and likely unnecesary
system.time({
  sas_connect_impl()
  sas_connect_impl()
})
#> SAS Connection established.
#> SAS Connection established.
#>    user  system elapsed
#>   0.140   0.068  34.758

sas_disconnect()
#> SAS Connection terminated.

# The new user-facing function doesn't re-connect
system.time({
  sas_connect()
  sas_connect()
})
#> SAS Connection established.
#> Already connected.
#>    user  system elapsed
#>   0.028   0.013  15.667

sas_disconnect()
#> SAS Connection terminated.

# The user can still request to reconnect on demand
system.time({
  sas_connect()
  sas_connect(reconnect = TRUE)
})
#> SAS Connection established.
#> SAS Connection established.
#>    user  system elapsed
#>   0.043   0.035  35.936
```

## Reference

* ml24. Link related funcitons via `#' @family` or `#' @seealso`. The goal is make it easier to discover and navigate related topics, like `sas_connect()` and `sas_disconnect()`. You may also consider documenting both in the same topic, e.g. by documenting `sas_connect()` as usual, then using `#' @rdname sas_connect` over the definition of `sas_disconnect()`. See https://roxygen2.r-lib.org/articles/reuse.html.

## Tests

* ml25. Answering your question

> I'm not sure if you have any recommendations as to how I can better test the `install_saspy()` and `configure_saspy()` functions.

* `install_saspy()`: I wouldn't worry too much. Based on its body it seems OK to just try it out on a fresh system, ideally in the three main platforms. It seems like a relatively thin wrapper around reticulate funcitons, which I expect to be pretty robust -- guessing from the quality of the developers and the size of the community that would report bugs.

* `configure_saspy()`: Yeah, it's hard to test "as is" because code that needs external resources is mixed with code that doesn't. Two approaches that can help:

1. [Use the humble object pattern]( https://martinfowler.com/bliki/HumbleObject.html):

> Some program elements are inherently difficult, or even impossible to test. Any logic in these elements is thus prone to bugs and difficult to evolve. To mitigate this problem, move as much as logic as possible out of the hard-to-test element and into other more friendly parts of the code base. By making untestable objects humble 1, we reduce the chances that they harbor evil bugs.

2. Mock pieces of code that get in the way of your testing the critical logic.

For example, `sas_engine()` is a fairly large function. A few lines of code really need a conneciton and get in the way of testing the main logic. We may use mocks to turn-off connection-checks, and to verify we've entered a specific `if` statement.

``` r
devtools::load_all()
#> ℹ Loading sasquatch
library(testthat)

# This is what head of sas_engine() looks like. The rest doesn't matter for
# this specific test.
sas_engine <- function (options) {
  # Some checks that force a connection even when we don't need it
  chk_connection()
  options$engine <- "txt"
  options$results <- "asis"

  code <- paste(options$code, collapse = "\n")

  # Some logic we want to test
  if (!isTRUE(getOption('knitr.in.progress'))) {
  r  eturn(sas_run_string(code))
  }

  "... a lot more code"
}

test_that("if the option `knitr.in.progress` is not TRUE returns early", {
  withr::local_options(knitr.in.progress = "not `TRUE`")

  testthat::local_mocked_bindings(
    # Turn off checks, since here we don't need a real connection
    chk_connection = function() NULL,
    # Test that knitr.in.progress forces sas_engine() to return early
    sas_run_string = function(code) "retuning early"
  )

  expect_equal(sas_engine(list(code = "whatever")), "retuning early")
})
#> Test passed 🎉
```

More resources:

* https://r-pkgs.org/testing-advanced.html
* https://books.ropensci.org/http-testing/
* https://webfakes.r-lib.org/

* ml26. Avoid dependencies between tests

```r
# Passes: Tests for sas_list() pass because test-connect() first calls sas_connect()
tests/testthat/
- test-connect()
- test-list()

# Fails: Tests for sas_list() fail because they run before test-connect.R calls sas_connect()
tests/testthat/
- test-aaa-list()
- test-connect()
```

You can avoid this dependency by calling `sas_connect()` in tests/testthat/test-helper.R. This runs once per test-suite rather than once per test or test file, resulting in a faster test-workflow and braking the dependency on the order of execution of the test-files.

* ml27. Avoid failures

`devtools::test()` reports 5 failing tests. That's only a few relative to the much larger number of successful tests, so I wouldn't worry too much about skipping them if it's hard to make them pass in the reviewer's systems.

```

ℹ Testing sasquatch
✔ | F W  S  OK | Context
✔ |          5 | connect [44.1s]
✔ |          4 | connection-checks [18.0s]
✖ | 1        1 | conversion
───────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-conversion.R:22:3): back and forth
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_equal(sas_from_r(df, "df"), df) at test-conversion.R:22:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
4. │   └─rlang::eval_bare(expr, quo_get_env(quo))
5. ├─sasquatch::sas_from_r(df, "df")
6. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
7. │ └─reticulate:::r_to_py.data.frame(x)
8. │   └─reticulate:::r_to_py_impl(x, convert = convert)
9. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
10. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
11.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
12.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)
───────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          3 | disconnect [21.2s]
✖ | 1       32 | engine [124.5s]
───────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-engine.R:193:3): non-html; default
<python.builtin.ModuleNotFoundError/python.builtin.ImportError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_module_import(module, convert = convert)`: ModuleNotFoundError: No module named 'markdownify'
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. └─sasquatch::sas_engine(options) at test-engine.R:193:3
2.   └─reticulate::import("markdownify") at sasquatch/R/engine.R:115:5
3.     └─reticulate:::py_module_import(module, convert = convert)
───────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          4 | file-copy [37.9s]
✔ |          4 | file-download [49.7s]
✔ |          2 | file-exists [31.7s]
✔ |          3 | file-remove [29.7s]
✔ |          3 | file-upload [20.8s]
✖ | 2        1 | from-r [10.3s]
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-from-r.R:24:3): R data.frame checks
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_warning(...) at test-from-r.R:24:3
2. │ └─testthat:::expect_condition_matching(...)
3. │   └─testthat:::quasi_capture(...)
4. │     ├─testthat (local) .capture(...)
5. │     │ └─base::withCallingHandlers(...)
6. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
7. ├─sasquatch::sas_from_r(df, "df")
8. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
9. │ └─reticulate:::r_to_py.data.frame(x)
10. │   └─reticulate:::r_to_py_impl(x, convert = convert)
11. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
13. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
14.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
15.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)

Error (test-from-r.R:43:3): Round trip
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_equal(sas_from_r(df, "df"), df) at test-from-r.R:43:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. ├─sasquatch::sas_from_r(df, "df")
5. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
6. │ └─reticulate:::r_to_py.data.frame(x)
7. │   └─reticulate:::r_to_py_impl(x, convert = convert)
8. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
9. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
10.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
11.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
✔ |          3 | list [6.8s]
✔ |          3 | run-file [14.7s]
✔ |          1 | run-string [5.6s]
✖ | 1        0 | to-r
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-to-r.R:71:3): SAS to R data.frame
Error in `sas_to_r("airline", "sashelp")`: object 'x' not found
Backtrace:
▆
1. └─sasquatch::sas_to_r("airline", "sashelp") at test-to-r.R:71:3
2.   └─reticulate::py_to_r(x) at sasquatch/R/to_r.R:40:3
3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
4.     └─reticulate:::py_to_r_cpp(x)
──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────

══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 415.3 s

── Failed tests ──────────────────────────────────────────────────────────────────────────────────────────────────────
Error (test-conversion.R:22:3): back and forth
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_equal(sas_from_r(df, "df"), df) at test-conversion.R:22:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. ├─sasquatch::sas_from_r(df, "df")
5. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
6. │ └─reticulate:::r_to_py.data.frame(x)
7. │   └─reticulate:::r_to_py_impl(x, convert = convert)
8. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
9. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
10.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
11.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)

Error (test-engine.R:193:3): non-html; default
<python.builtin.ModuleNotFoundError/python.builtin.ImportError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_module_import(module, convert = convert)`: ModuleNotFoundError: No module named 'markdownify'
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. └─sasquatch::sas_engine(options) at test-engine.R:193:3
2.   └─reticulate::import("markdownify") at sasquatch/R/engine.R:115:5
3.     └─reticulate:::py_module_import(module, convert = convert)

Error (test-from-r.R:24:3): R data.frame checks
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_warning(...) at test-from-r.R:24:3
2. │ └─testthat:::expect_condition_matching(...)
3. │   └─testthat:::quasi_capture(...)
4. │     ├─testthat (local) .capture(...)
5. │     │ └─base::withCallingHandlers(...)
6. │     └─rlang::eval_bare(quo_get_expr(.quo), quo_get_env(.quo))
7. ├─sasquatch::sas_from_r(df, "df")
8. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
9. │ └─reticulate:::r_to_py.data.frame(x)
10. │   └─reticulate:::r_to_py_impl(x, convert = convert)
11. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
12. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
13.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
14.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)

Error (test-from-r.R:43:3): Round trip
<python.builtin.TypeError/python.builtin.Exception/python.builtin.BaseException/python.builtin.object/error/condition>
Error in `py_call_impl(callable, call_args$unnamed, call_args$named)`: TypeError: 'list' object cannot be interpreted as an integer
Run `reticulate::py_last_error()` for details.
Backtrace:
▆
1. ├─testthat::expect_equal(sas_from_r(df, "df"), df) at test-from-r.R:43:3
2. │ └─testthat::quasi_label(enquo(object), label, arg = "object")
3. │   └─rlang::eval_bare(expr, quo_get_env(quo))
4. ├─sasquatch::sas_from_r(df, "df")
5. │ ├─reticulate::r_to_py(x) at sasquatch/R/from_r.R:55:3
6. │ └─reticulate:::r_to_py.data.frame(x)
7. │   └─reticulate:::r_to_py_impl(x, convert = convert)
8. ├─reticulate (local) `<fn>`(`<dttm>`, FALSE)
9. └─reticulate:::r_to_py.POSIXt(`<dttm>`, FALSE)
10.   └─datetime$datetime$fromtimestamp(as.double(x), tz = tz)
11.     └─reticulate:::py_call_impl(callable, call_args$unnamed, call_args$named)

Error (test-to-r.R:71:3): SAS to R data.frame
Error in `sas_to_r("airline", "sashelp")`: object 'x' not found
Backtrace:
▆
1. └─sasquatch::sas_to_r("airline", "sashelp") at test-to-r.R:71:3
2.   └─reticulate::py_to_r(x) at sasquatch/R/to_r.R:40:3
3.     ├─reticulate::is_py_object(x <- py_to_r_cpp(x))
4.     └─reticulate:::py_to_r_cpp(x)

[ FAIL 5 | WARN 0 | SKIP 0 | PASS 69 ]
```

* ml28. Consider skipping tests when there's no connection. For example, this failure IS informative but currently the relevant error message is somewhat burried by the less useful backtrace.

``` r
devtools::load_all()
#> ℹ Loading sasquatch

# tests/testthat/helper.R
skip_if_unconnected <- function() {
  skip_if(
  is.null(sas_get_session()),
  "No active SAS session. Use sas_connect() to start one."
  )
}

# tests/testthat/test-*.R
testthat::test_that("needs a sas connection", {
  skip_if_unconnected()
  sas_list()
})
#> ── Skip: needs a sas connection ────────────────────────────────────────────────
#> Reason: No active SAS session. Use sas_connect() to start one.

# tests/testthat/test-*.R
# This IS informative but:
# * Doesn't communicate the develper the the connection is a requirement
# * Burries the relevant information with irrelevant clutter from the backtrace
testthat::test_that("needs a sas connection", {
  sas_list()
})
#> ── Error: needs a sas connection ───────────────────────────────────────────────
#> <chk_error/rlang_error/error/condition>
#> Error in `sas_list()`: No active SAS session. Use sas_connect() to start one.
#> Backtrace:
#>     ▆
#>  1. └─sasquatch::sas_list()
#>  2.   └─sasquatch:::chk_connection() at sasquatch/R/list.R:17:3
#>  3.     └─chk::abort_chk("No active SAS session. Use sas_connect() to start one.") at sasquatch/R/connection-checks.R:15:3
#>  4.       └─chk::err(..., n = n, tidy = tidy, class = "chk_error", call = call)
#>  5.         └─rlang::abort(msg, class = class, !!!args[named], call = call)
#> Error:
#> ! Test failed
```

* ml29. If you need to create then destroy a connection multiple times, consider a helper `local_sas_connect()` to create self-destroying connection.

``` r
devtools::load_all()
#> ℹ Loading sasquatch

[…]
