### 2025-02-02 #673 @maurolepore
Dear @ryanzomorrodi FYI my EiC rotation starts today and I'm reviewing all open issues. I see I reached out about 3 weeks ago, but got no response. Are you still planning to continue?

If you can't continue right now you can choose to put the submission on hold. This is totally fine as we understand things can change -- so no pressure.

---

https://devguide.ropensci.org/softwarereview_policies.html#policiesreviewprocess

> * The author can choose to have their submission put on hold (editor applies the holding label). The holding status will be revisited every 3 months, and after one year the issue will be closed.

> * If the author hasn’t requested a holding label, but is simply not responding, we should close the issue within one month after the last contact intent. This intent will include a comment tagging the author, but also an email using the email address listed in the DESCRIPTION of the package which is one of the rare cases where the editor will try to contact the author by email.
