### 2025-02-04 #673 @ryanzomorrodi
Hi @maurolepore ,

Sorry for the long delay. No need to put a hold. I have been working on sasquatch, mostly just playing with the sas_engine for non html formats, but I haven't really arrived at a great solution.

I introduced another python dependency (markdownify), which explains one of the errors you got, but weirdly not all. I'm very confused by the R to SAS conversion errors, but all this is to say that I will be addressing everything by the end of the week.
