### 2025-02-10 a909722
m20: Add SAS ODA Config to README
