+++ README.Rmd
@@ -63,7 +63,7 @@ currently implemented within `sasquatch` like latex pdfs or non-HTML5 HTML.
 
 `sasquatch` may be beneficial to you if you...  
 
-* Rely on remote SAS client  
+* Rely on a remote SAS client  
 * Desire interactive SAS functionality while developing  
 * Require remote SAS file management  
 * Would like to be able to easily send data back and forth between SAS and R   
