### 2025-02-10 #673 @ryanzomorrodi
I am still working on things, but I want to write down a few of my thoughts so I don't forget them later, so I will edit this as I address more of the questions.

m19: That sounds very logical. I have reconfigured the order of README so that it fits that natural sequence.

m20: I also added a SAS ODA configuration explanation to the readme for users who just want to get up and running quickly.

m21: I actually did mean RStudio on that one. The gif below that sentence tries to depict what I mean, but in RStudio, if you press the green play button on a chunk, it will not send the output of the chunk to the console/plots/viewer, but rather embed the output within your view of the document itself.

![Image](https://github.com/user-attachments/assets/79400dc4-d08c-46f4-8d26-93d8ca748789)

VSCode, Positron, and neovim don't have this capability (at least to my knowledge, but its definitely not a default). In fact in Positron or VSCode, you can only run chunks that the IDE already knows about so, even if you set an engine in knitr, the IDE won't give you the option to even run it. (Hence why there's an example keybinding for sas_run_selected for Positron).

m23: The connection to SAS is async, which unfortunately means the connection can be disrupted without the user knowing. The only way to truly know if the session is active is actually try to send something. This is why there is a `execute_if_connection_active()` around every time a session command is made.

The function allows me to
- Run the session command
- Check if the session connection has been broken
- If it has, then give an informative error. Else, just return whatever was supposed to be returned anyway.

This has implications for `sas_connect()` as well because if I were to implement a `reconnect` parameter and a connection has been unknowingly severed, then `sas_connect()` will not reconnect you unless you pass in `reconnect = TRUE` and will tell you that it's already connected. The only way to mitigate this call a session function that doesn't alter the session and check after the fact if the session has dropped, but this would make sas_connect() function take even longer to run (its already by far the slowest function).
