### 2025-02-11 158b2ea
m24: Link related functions
