+++ R/file-remove.R
@@ -9,6 +9,7 @@
 #' 
 #' @export
 #' 
+#' @family file management functions
 #' @examples
 #' \dontrun{
 #' # connect to SAS
