+++ R/disconnect.R
@@ -7,6 +7,7 @@
 #' 
 #' @export
 #' 
+#' @family session management functions
 #' @examples
 #' \dontrun{
 #' sas_disconnect()
