+++ R/connect.R
@@ -16,6 +16,7 @@
 #' 
 #' @export
 #' 
+#' @family session management functions
 #' @examples
 #' \dontrun{
 #' sas_connect(cfgname = "oda")
