### 2025-02-11 6e1cf24
Refactor of configure_saspy()
