+++ R/saspy-configure.R
@@ -5,6 +5,46 @@
 #' according to a specified template.
 #' 
 #' @param template Default template to base configuration files off of.
+#' @param overwite Can new configuration files overwrite existing config files (if they exist)?
+#' 
+#' @details
+#' Configuration for SAS can vary greatly based on your computer's operating 
+#' system and the SAS platform you wish to connect to (see 
+#' `vignette("configuration")` for more information). 
+#' 
+#' Regardless of your desired configuration, configuration always starts with
+#' the creation of a `sascfg_personal.py` file within the `SASPy` package
+#' installation. This will look like:
+#' 
+#' ```
+#' SAS_config_names = ['config_name']
+#' 
+#' config_name = {
+#' 
+#' }
+#' ```
+#' 
+#' `SAS_config_names` should contain a string list of the variable names 
+#' of all configurations. Configurations are specified as dictionaries,
+#' and configuration parameters depend on the access method.
+#' 
+#' Additionally, some access methods will require an additional 
+#' authentication file (`.authinfo` for Linux and Mac, `_authinfo` 
+#' for Windows) stored in the user's home directory, which are
+#' constructed as follows:
+#' 
+#' ```
+#' config_name user {your username} password {your password}
+#' ```
+#' 
+#' ### Templates
+#' 
+#' The `"none"` template simply creates a `sascfg_personal.py` file within 
+#' the `SASPy` package installation. 
+#' 
+#' The `"oda"` template will set up a configuration for SAS On Demand for
+#' Academics. The `sascfg_personal.py` and `authinfo` files will be
+#' automatically configured using the information you provide through prompts.
 #' 
 #' @return No return value.
 #' 
@@ -16,115 +56,50 @@
 #' config_saspy()
 #' }
 configure_saspy <- function(
-  template = c("none", "oda")
+  template = c("none", "oda"),
+  overwrite = FALSE
 ) {
-  python <- reticulate::py_discover_config("saspy", "r-saspy")
-  saspy_path <- python$required_module_path
-  sascfg_personal_path <- paste0(saspy_path, "/sascfg_personal.py")
-
-  config_name <- "config_name"
-  config <- list()
-
-  if (template[1] == "oda") {
-    config_name <- "oda"
+  if (template[1] == "none") {
+    configs <- list(config_name = list())
+  } else if (template == "oda") {
     oda_username <- readline("Enter your ODA username: ")
     oda_password <- readline("Enter your ODA password: ")
-    authinfo <- paste("oda user", oda_username, "password", oda_password)
+    write_authinfo("oda", oda_username, oda_password, overwrite)
 
-    if(.Platform$OS.type == "windows") {
-      authinfo_path <- paste0(get_home_dir(), "/_authinfo")
-    } else {
-      authinfo_path <- paste0(get_home_dir(), "/.authinfo")
-    }
-
-    if (file.exists(authinfo_path)) {
-      overwrite_auth = menu(
-        c("Yes", "No"), 
-        title = "The authinfo file already exists and you're about to overwrite it.\nDo you want to proceed anyway?"
-      )
-      if (overwrite_auth == 0) {
-        stop("The authinfo file already exists.")
-      }
-    }
-    if (!exists("overwrite_auth") || overwrite_auth == 1) {
-      cat("Writing to", authinfo_path, "\n")
-      cat(authinfo, file = authinfo_path)
[…]
