+++ README.Rmd
@@ -31,11 +31,11 @@ to create reproducible multilingual reports. `sasquatch` can:
 * Conduct basic file management on a SAS client
 * Render SAS output within quarto documents.
 
-`sasquatch` relies on the [`SASPy`](https://sassoftware.github.io/saspy/) Python
-package and the `reticulate` R package to interoperate with Python. Check out 
-`vignette("configuration")` for guidance on `SASPy` configuration.
-
 ### Comparison with similar packages
+
+`sasquatch` relies on the [`SASPy`](https://sassoftware.github.io/saspy/) Python
+package and the `reticulate` R package to interoperate with Python. There exist
+similar packages, which work similarly to achieve related goals.
 
 [`sasr`](https://CRAN.R-project.org/package=sasr)
 
@@ -49,6 +49,7 @@ or quarto functionality.
 but it primarily focuses on solely on `knitr` engine support.
 * The `configSAS` engine HTML output CSS styles interfere with the rest of the
 document and SAS code output is not contained within a code block.
+* `configSAS` is not actively maintained.
 
 [`SASmarkdown`](https://CRAN.R-project.org/package=SASmarkdown)
 
@@ -69,7 +70,7 @@ currently implemented within `sasquatch` like latex pdfs or non-HTML5 HTML.
 * Would like to be able to easily send data back and forth between SAS and R   
 without the use of intermediate files  
 
-If you require pdf `knitr` support and have a local installation of SAS, I would
+If you require pdf `knitr` engine support and have a local installation of SAS, I would
 recommend using `SASmarkdown`at this time.
 
 ## Installation
