### 2025-02-11 64cec83
Adding testing to configure_saspy()
