### 2025-02-11 401377a
Simplify tests with local tempfile bindings
