+++ tests/testthat/test-run-file.R
@@ -1,28 +1,24 @@
 test_that("generate SAS widget from file", {
+  sas_connect()
   skip_on_cran()
   skip_if_offline()
-  withr::defer(file.remove(local_path))
-  
-  local_path <- tempfile(pattern="temp", fileext=".sas")
+  local_path <- withr::local_tempfile(pattern = "temp", fileext = ".sas", lines = "PROC MEANS DATA = sashelp.cars; RUN;")
 
   "widget if no output path specified"
-  cat("PROC MEANS DATA = sashelp.cars; RUN;", file = local_path)
   expect_s3_class(sas_run_file(local_path), c("sas_widget", "htmlwidget"))
 })
 
 test_that("generate output html and log from file", {
   skip_on_cran()
   skip_if_offline()
-  withr::defer(file.remove(local_path))
-  withr::defer(file.remove(local_html))
-  withr::defer(file.remove(local_log))
+  local_dir_path <- withr::local_tempdir(pattern = "temp")
   
-  local_path <- tempfile(pattern="temp", fileext=".sas")
+  local_path <- paste0(local_dir_path, "/", basename(tempfile(pattern = "temp", fileext = ".sas")))
+  cat("PROC MEANS DATA = sashelp.cars; RUN;", file = local_path)
   local_html <- sub(".sas", ".html", local_path, fixed = TRUE)
   local_log <- sub(".sas", ".log", local_path, fixed = TRUE)
 
   "generate html if no output path specified"
-  cat("PROC MEANS DATA = sashelp.cars; RUN;", file = local_path)
   sas_run_file(local_path, local_html)
   expect_true(file.exists(local_html))
 })
@@ -30,17 +26,15 @@ test_that("generate output html and log from file", {
 test_that("overwrite output html and log from file", {
   skip_on_cran()
   skip_if_offline()
-  withr::defer(file.remove(local_path))
-  withr::defer(file.remove(local_html))
-  withr::defer(file.remove(local_log))
+  local_dir_path <- withr::local_tempdir(pattern = "temp")
   
-  local_path <- tempfile(pattern="temp", fileext=".sas")
+  local_path <- paste0(local_dir_path, "/", basename(tempfile(pattern = "temp", fileext = ".sas")))
+  cat("PROC MEANS DATA = sashelp.cars; RUN;", file = local_path)
   local_html <- sub(".sas", ".html", local_path, fixed = TRUE)
+  cat("test", file = local_html)
   local_log <- sub(".sas", ".log", local_path, fixed = TRUE)
 
   "don't overwrite"
-  cat("PROC MEANS DATA = sashelp.cars; RUN;", file = local_path)
-  cat("test", file = local_html)
   expect_error(sas_run_file(local_path, local_html), "already exists. If you would like to overwrite the file, use overwrite = TRUE.", fixed = TRUE)
 
   "overwrite output"
