### 2025-02-11 a143747
Don't skip configuration tests if offline
