### 2025-02-12 d7cd438
Remove all non-html knitr support
