+++ man/sas_set_tempdir.Rd
@@ -1,23 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/engine.R
-\name{sas_set_tempdir}
-\alias{sas_set_tempdir}
-\title{Set SAS temporary directory}
-\usage{
-sas_set_tempdir(path)
-}
-\arguments{
-\item{path}{string; Path for SAS to write intermediate files to for the
-\code{knitr} \code{sas_engine()}.}
-}
-\value{
-No return value.
-}
-\description{
-In order to create markdown output for non-HTML formats, SAS must render
-intermediate files to a directory on the SAS. This function allows you
-to set what directory intermediate files are writen to.
-}
-\seealso{
-\code{\link[=sas_engine]{sas_engine()}}
-}
