+++ NAMESPACE
@@ -16,6 +16,5 @@ export(sas_list)
 export(sas_run_file)
 export(sas_run_selected)
 export(sas_run_string)
-export(sas_set_tempdir)
 export(sas_to_r)
 importFrom(knitr,is_html_output)
