### 2025-02-12 4871753
Add a memoized sas_connect
