+++ tests/testthat/test-file-exists.R
@@ -1,6 +1,7 @@
 test_that("checking if file exists on SAS", {
   skip_on_cran()
   skip_if_offline()
+  sas_connect_if_no_session()
   local_path <- withr::local_tempfile(pattern = "temp", fileext = ".sas", lines = "PROC MEANS DATA = sashelp.cars; RUN;")
   withr::defer(sas_file_remove(sas_path))
   
