+++ tests/testthat/test-connect.R
@@ -1,7 +1,7 @@
 test_that("default connection", {
   skip_on_cran()
   skip_if_offline()
-  withr::defer(suppressMessages(sas_disconnect()))
+  try(suppressMessages(sas_disconnect()), silent = TRUE)
 
   expect_message(sas_connect(), "SAS Connection established.", fixed = TRUE)
   expect_s3_class(sas_get_session(), c("saspy.sasbase.SASsession", "python.builtin.object"))
@@ -10,6 +10,7 @@ test_that("default connection", {
 test_that("specified cfgname", {
   skip_on_cran()
   skip_if_offline()
+  try(supressMessages(sas_disconnect()), silent = TRUE)
 
   "non-string cfgname"
   expect_error(sas_connect(1), "must be a string")
