+++ tests/testthat/test-file-copy.R
@@ -1,6 +1,7 @@
 test_that("copying file from SAS", {
   skip_on_cran()
   skip_if_offline()
+  sas_connect_if_no_session()
   local_path <- withr::local_tempfile(pattern = "temp", fileext = ".sas", lines = "PROC MEANS DATA = sashelp.cars; RUN;")
   withr::defer(sas_file_remove(sas_path))
   withr::defer(sas_file_remove(sas_copy_path))
