+++ pkgdown.yml
@@ -3,8 +3,8 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   configuration: configuration.html
-  sasquatch: sasquatch.html
-last_built: 2024-12-30T07:49Z
+  overview: overview.html
+last_built: 2025-02-12T08:34Z
 urls:
   reference: https://ryanzomorrodi.github.io/sasquatch/reference
   article: https://ryanzomorrodi.github.io/sasquatch/articles
