### 2025-02-13 ddea80e
Convert factors to characters before reticulate (#5)
