### 2025-03-04 #673 @maurolepore
Hi @ryanzomorrodi, just noticing a bit of inactivity so pinging you to ask for an update. Remember our options are basically 1) take the package to a reviewer-friendly level now or 2) put the review on hold and check-in for progress every 3 months for up to  year.
