### 2025-03-10 #673 @maurolepore
Sure, no pressure. The bot will touch base every 3 months for a year.

 All the best with your main responsibilities.
