+++ tests/testthat/test-configure.R
@@ -7,7 +7,10 @@ test_that("oda configuration - sascfg_personal", {
   configs <- list(
     oda = list(
       java = "usr/bin/java",
-      iomhost = list('odaws01-usw2-2.oda.sas.com', 'odaws02-usw2-2.oda.sas.com'),
+      iomhost = list(
+        'odaws01-usw2-2.oda.sas.com',
+        'odaws02-usw2-2.oda.sas.com'
+      ),
       iomport = 8591L,
       encoding = "utf-8",
       authkey = "oda"
@@ -22,7 +25,7 @@ test_that("oda configuration - sascfg_personal", {
   expect_equal(
     readLines(tempfile),
     c(
-      "SAS_config_names = ['oda']", 
+      "SAS_config_names = ['oda']",
       "oda = {'java': 'usr/bin/java', 'iomhost': ['odaws01-usw2-2.oda.sas.com', 'odaws02-usw2-2.oda.sas.com'], 'iomport': 8591, 'encoding': 'utf-8', 'authkey': 'oda'}"
     )
   )

[…]
