### 2025-03-13 68f9aaf
replace chk for cli
