### 2025-03-30 ad911e3
Adding cfgname check to abort when cfgname not found
