### 2025-03-30 4c5f5c4
Remove sas_disconnect connection check. Do not alter session object.
