### 2025-03-30 fbb1de9
Use mock-able binding for `sas_list()` in `sas_file_exists()``
