### 2025-03-30 85542fc
Make connection checks more informative
