### 2025-03-30 1481a8f
Make error messages more informative for `sas_file_download()`
