### 2025-03-30 d380418
Remove all timezones when converting to sas to make roundtrip more predictable
