### 2025-03-30 bb2ffb2
Fix `configure_saspy()` menu breaking error
