### 2025-03-30 911c1b6
Remove timezones when converting to R to make round trip more predictable
