### 2025-03-30 1699263
Completely revamp testing installation and configuration requirements and add snapshot testing
