+++ tests/testthat/test-from-r.R
@@ -1,67 +1,269 @@
-test_that("R data.frame checks", {
-  skip_on_cran()
-  skip_if_offline()
-  sas_connect_if_no_session()
-  withr::defer(sas_get_session()$submit(
-    "proc datasets library=WORK;delete df;run;"
-  ))
+# processing tests
+test_that("double should not be altered", {
+  x <- data.frame(a = runif(1000, min = -1, max = 1))
 
-  df_all <- data.frame(
-    a = c(1, 2.5, NA),
-    b = c(1:2, NA),
-    c = c(T, F, NA),
-    d = c("a", "b", NA),
-    e = factor(c("a", "b", NA)),
-    f = as.Date("2015-12-09") + c(1:2, NA),
-    g = as.POSIXct("2015-12-09 10:51:34.5678", tz = "UTC") + c(1:2, NA),
-    h = I(as.list(c(1:2, NA))),
-    i = I(list(list(1, 2:3), list(4:6), list(NA)))
+  expect_equal(x, from_r_processing(x)$x)
+
+  x$a[sample(1:nrow(x), 100)] <- NA
+
+  expect_equal(from_r_processing(x)$x, x)
+})
+
+test_that("integer should become a double", {
+  x <- data.frame(a = sample(-1000:1000, 1000, replace = TRUE))
+  x_expected <- x
+  x_expected$a <- as.double(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+
+  x$a[sample(1:nrow(x), 100)] <- NA
+  x_expected <- x
+  x_expected$a <- as.double(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+})
+
+test_that("logical should become a double", {
+  x <- data.frame(a = sample(c(TRUE, FALSE), 1000, replace = TRUE))
+  x_expected <- x
+  x_expected$a <- as.double(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+
+  x$a[sample(1:nrow(x), 100)] <- NA
+  x_expected <- x
+  x_expected$a <- as.double(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+})
+
+test_that("character should not be altered", {
+  x <- data.frame(
+    a = sample(c("apple", "pear", "orange", "cherry"), 1000, replace = TRUE)
   )
 
-  "list columns error"
-  expect_error(
-    sas_from_r(df_all, "df"),
-    "must only contain logical, integer, numeric, factor, character, POSIXct, or Date columns",
-    fixed = TRUE
+  expect_equal(from_r_processing(x)$x, x)
+
+  x$a[sample(1:nrow(x), 100)] <- NA
+
+  expect_equal(from_r_processing(x)$x, x)
+})
+
+test_that("factor should become a character", {
+  x <- data.frame(
+    a = as.factor(sample(
+      c("apple", "pear", "orange", "cherry"),
+      1000,
+      replace = TRUE
+    ))
+  )
+  x_expected <- x
+  x_expected$a <- as.character(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+
+  x$a[sample(1:nrow(x), 100)] <- NA
+  x_expected <- x
+  x_expected$a <- as.character(x_expected$a)
+
+  expect_equal(from_r_processing(x)$x, x_expected)
+})
+
+test_that("POSIXct should be converted to UTC", {
+  x <- data.frame(
+    a = as.POSIXct(format(
+      Sys.time() + sample(-1000:1000, 1000, replace = TRUE)
+    ))
+  )
[…]
