### 2025-03-30 10ab927
Fix empty vector list check
