+++ tests/testthat/test-list.R
@@ -6,7 +6,7 @@ test_that("empty vector returned for path that doesn't exist", {
 
   file_list <- sas_list("~/thisIsARandomPathThatDoesntExist")
 
-  expect_vector(is.character(file_list))
+  expect_true(is.character(file_list))
   expect_length(file_list, 0)
 })
 
