+++ man/sas_from_r.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/from_r.R
+% Please edit documentation in R/from-r.R
 \name{sas_from_r}
 \alias{sas_from_r}
 \title{Convert R table to SAS}
@@ -30,7 +30,7 @@ as follows:
 \item double -> numeric
 \item factor -> character
 \item character -> character
-\item POSIXct -> numeric (datetime)
+\item POSIXct -> numeric (datetime; timezones are lost)
 \item Date -> numeric (date)
 }
 }
