+++ man/sas_to_r.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/to_r.R
+% Please edit documentation in R/to-r.R
 \name{sas_to_r}
 \alias{sas_to_r}
 \title{Convert SAS table to R}
@@ -28,7 +28,7 @@ as follows:
 }
 
 In the conversion process dates and datetimes are converted to local
-time. If utilizing another timezone, use \code{as.POSIXct()} or
+time. If utilizing another timezone, use \verb{attr(date, "tzone") <-} or
 \code{lubridate::with_tz()} to convert back to the desired time zone.
 }
 \examples{
