+++ R/sasquatch-package.R
@@ -2,6 +2,6 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom knitr is_html_output
+#' @importFrom knitr is_html_output is_latex_output
 ## usethis namespace: end
 NULL
