+++ R/engine.R
@@ -59,6 +59,10 @@ sas_engine <- function(options) {
       log <- wrap_in_pre(results$LOG)
       out <- wrap_in_panel_tabset(lst, log)
     }
+  } else if (is_latex_output()) {
+    execute_if_connection_active(
+      out <- .sas_run_latex(code)
+    )
   } else {
     cli::cli_abort(
       "{.fun sasquatch::sas_engine} cannot produce non-html output."
@@ -109,4 +113,69 @@ wrap_in_panel_tabset <- function(lst, log) {
     ',
     sep = "\n"
   )
+}
+
+.sas_run_latex <- function(code) {
+  sas_path <- .pkgenv$session$`_io`$`_sb`$workpath
+  output_tex <- basename(tempfile(pattern = "", tmpdir = ".", fileext = ".tex"))
+
+  wrapped_code <- paste(
+    "ods noproctitle;",
+    "ods listing close;",
+    paste0(
+      "ods tagsets.simplelatex path = '",
+      sas_path,
+      "' gpath = '",
+      sas_path,
+      "' file='",
+      output_tex,
+      "' (no_top_matter no_bottom_matter) stylesheet='sas_style.sty'(url='sas');"
+    ),
+    "ods graphics / imagename='glabel' outputfmt=png;",
+    "OPTIONS NONUMBER NODATE PAGESIZE = MAX FORMCHAR = '|----|+|---+=|-/<>*' FORMDLIM=' ';title;",
+    code,
+    "ods tagsets.simplelatex close;",
+    sep = "\n"
+  )
+
+  .sas_run_string(wrapped_code)
+
+  workspace_files <- .sas_list(sas_path)
+  output_files <- grep("(tex|png|sty)$", workspace_files, value = TRUE)
+
+  lapply(
+    output_files,
+    \(file)
+      .sas_file_download(
+        file.path(".sasquatch", file),
+        paste0(sas_path, "/", file)
+      )
+  )
+  lapply(
+    output_files,
+    \(file) .sas_file_remove(paste0(sas_path, "/", file))
+  )
+
+  # replace problematic styles
+  sty <- readLines(".sasquatch/sas_style.sty")
+  sty2 <- sub("\\newcommand{\\color}[2][]{}", "", sty, fixed = TRUE)
+  sty3 <- sub(
+    "\\newcommand{\\sascaption}[2][l]{\\marginpar[#1]{\\fbox{\\parbox{0.7in}{\\sasScaption{#2}}}}}",
+    "\\newcommand{\\sascaption}[2][l]{\\reversemarginpar{\\fbox{\\parbox{0.9in}{\\sasScaption{#2}}}}}",
+    sty2,
+    fixed = TRUE
+  )
+  writeLines(sty3, ".sasquatch/sas_style.sty")
+
+  tex_file <- grep("tex$", output_files, value = TRUE)
+  output_tex <- readLines(file.path(".sasquatch", tex_file))
+  graphics_lines <- grepl("\\includegraphics{", output_tex, fixed = TRUE)
+  output_tex[graphics_lines] <- gsub(
+    sas_path,
+    ".sasquatch/",
+    output_tex[graphics_lines],
+    fixed = TRUE
+  )
+
+  gsub("\\pagebreak", " ", output_tex)
 }
