+++ NAMESPACE
@@ -18,3 +18,4 @@ export(sas_run_selected)
 export(sas_run_string)
 export(sas_to_r)
 importFrom(knitr,is_html_output)
+importFrom(knitr,is_latex_output)
