### 2025-03-31 2baf7c4
delete tex and graphics after insertion
