### 2025-04-01 ca842c4
Use reticulate::py_require()
