### 2025-04-01 cf83707
Fix tests when oda has not been configured
