+++ tests/testthat/test-connect.R
@@ -6,18 +6,10 @@ test_that("non-existing cfgname throws error", {
 
   expect_snapshot(
     sas_connect(cfgname = "some config that doesn't exist"),
-    transform = function(lines) {
-      avail_config <- "Available configurations include: "
-      gsub(paste0(avail_config, "(.*)"), avail_config, lines)
-    },
     error = TRUE
   )
   expect_snapshot(
     sas_connect(cfgname = "anotherconfigthatdoesntexist"),
-    transform = function(lines) {
-      avail_config <- "Available configurations include: "
-      gsub(paste0(avail_config, "(.*)"), avail_config, lines)
-    },
     error = TRUE
   )
 })
@@ -26,7 +18,6 @@ test_that("existing cfgname establishes connection", {
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
-  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   expect_message(
@@ -40,11 +31,51 @@ test_that("existing cfgname establishes connection", {
   )
 })
 
+test_that("reconnecting warns user if `reconnect = FALSE` and doesn't replace connection", {
+  skip_on_cran()
+  skip_if_offline()
+  skip_if_no_saspy_install()
+  suppressMessages(sas_disconnect())
+
+  suppressMessages(sas_connect(cfgname = "oda"))
+
+  config <- sas_get_session()
+
+  expect_warning(
+    sas_connect(cfgname = "oda"),
+    "SAS connection already established. Specify `reconnect = TRUE` to establish a new connection.",
+    fixed = TRUE
+  )
+
+  expect_equal(sas_get_session(), config)
+})
+
+test_that("reconnecting establishes a new connection if `reconnect = TRUE`", {
+  skip_on_cran()
+  skip_if_offline()
+  skip_if_no_saspy_install()
+  suppressMessages(sas_disconnect())
+
+  suppressMessages(sas_connect(cfgname = "oda"))
+
+  config <- sas_get_session()
+
+  suppressMessages(sas_connect(cfgname = "oda", reconnect = TRUE))
+
+  expect_s3_class(
+    .pkgenv$session,
+    c("saspy.sasbase.SASsession", "python.builtin.object")
+  )
+
+  expect_failure(
+    expect_equal(sas_get_session(), config)
+  )
+})
+
 test_that("default connection establishes connection", {
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
-  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   expect_message(sas_connect(), "SAS connection established.", fixed = TRUE)
