+++ tests/testthat/_snaps/connect.md
@@ -6,7 +6,7 @@
       Error in `check_cfgname()`:
       ! `cfgname` must specify an existing configuration.
       x `"some config that doesn't exist" cannot be found.
-      i Available configurations include: 
+      i Available configurations include: "oda"
 
 ---
 
@@ -16,5 +16,5 @@
       Error in `check_cfgname()`:
       ! `cfgname` must specify an existing configuration.
       x `"anotherconfigthatdoesntexist" cannot be found.
-      i Available configurations include: 
+      i Available configurations include: "oda"
 
