+++ tests/testthat/test-connect.R
@@ -6,10 +6,18 @@ test_that("non-existing cfgname throws error", {
 
   expect_snapshot(
     sas_connect(cfgname = "some config that doesn't exist"),
+    transform = function(lines) {
+      avail_config <- "Available configurations include: "
+      gsub(paste0(avail_config, "(.*)"), avail_config, lines)
+    },
     error = TRUE
   )
   expect_snapshot(
     sas_connect(cfgname = "anotherconfigthatdoesntexist"),
+    transform = function(lines) {
+      avail_config <- "Available configurations include: "
+      gsub(paste0(avail_config, "(.*)"), avail_config, lines)
+    },
     error = TRUE
   )
 })
@@ -18,6 +26,7 @@ test_that("existing cfgname establishes connection", {
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
+  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   expect_message(
@@ -35,6 +44,7 @@ test_that("reconnecting warns user if `reconnect = FALSE` and doesn't replace co
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
+  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   suppressMessages(sas_connect(cfgname = "oda"))
@@ -54,6 +64,7 @@ test_that("reconnecting establishes a new connection if `reconnect = TRUE`", {
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
+  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   suppressMessages(sas_connect(cfgname = "oda"))
@@ -76,6 +87,7 @@ test_that("default connection establishes connection", {
   skip_on_cran()
   skip_if_offline()
   skip_if_no_saspy_install()
+  skip_if_no_configuration("oda")
   suppressMessages(sas_disconnect())
 
   expect_message(sas_connect(), "SAS connection established.", fixed = TRUE)
