### 2025-04-09 dfa6441
delete example
