+++ example/test.qmd
@@ -1,23 +1,0 @@
----
-title: "sasquatch pdf demo"
-format: pdf
----
-
-```{r}
-devtools::load_all()
-sas_connect()
-```
-
-```{sas}
-proc print data = sashelp.cars(obs=10);
-run;
-```
-
-```{sas}
-proc model data=sashelp.cars;
-    parms b0 b1 b2;
-    MSRP = b0 + b1*Horsepower + b2*Weight;
-    fit MSRP / outest=estimates;
-run;
-```
-
