### 2025-04-10 #673 @mpadge
@ryanzomorrodi It's not `#nocov` statements, it's because we don't, and cannot, use your full test configuration you have on GitHub runners. In such cases, we're happy to rely on the more accurate codecov statistics. @maurolepore Please ignore the falling check - coverage here is good.
