+++ inst/htmlwidgets/sas_widget.js
@@ -4,25 +4,25 @@ HTMLWidgets.widget({
 
   type: 'output',
 
-  factory: function(el, width, height) {
+  factory: function (el, width, height) {
 
     // TODO: define shared variables for this instance
 
     return {
 
-      renderValue: function(x) {
+      renderValue: function (x) {
         let lst = x.lst;
         let log = x.log;
-        
+
 
         el.innerHTML = `
 <style>body {overflow: auto !important;}</style>
 <ul class="nav nav-tabs" id="myTab" role="tablist">
   <li class="nav-item" role="presentation">
-    <button class="nav-link active" id="lst-tab" data-bs-toggle="tab" data-bs-target="#lst" type="button" role="tab" aria-controls="lst" aria-selected="true">lst</button>
+    <button class="nav-link active" id="lst-tab" data-bs-toggle="tab" data-bs-target="#lst" type="button" role="tab" aria-controls="lst" aria-selected="true">Listing</button>
   </li>
   <li class="nav-item" role="presentation">
-    <button class="nav-link" id="log-tab" data-bs-toggle="tab" data-bs-target="#log" type="button" role="tab" aria-controls="log" aria-selected="false">log</button>
+    <button class="nav-link" id="log-tab" data-bs-toggle="tab" data-bs-target="#log" type="button" role="tab" aria-controls="log" aria-selected="false">Log</button>
   </li>
 </ul>
 <div style = "height: calc(100% - 49px) !important" class="tab-content" id="myTabContent">
@@ -34,8 +34,8 @@ HTMLWidgets.widget({
 
       },
 
-      resize: function(width, height) {
-        
+      resize: function (width, height) {
+
       }
 
     };
