+++ NEWS.md
@@ -1,0 +1,3 @@
+# sasquatch (development version)
+
+## rOpenSci Review Changes
