### 2025-04-15 ac07b50
Adjust htmlwidget size based off of the content
