### 2025-04-27 #673 @thisisnic
Thanks for inviting me to be involved - my review is pasted below.  Hope this is useful @ryanzomorrodi , and happy for any meta-feedback on the review itself @maurolepore!

---

I did this review without reading any previous context above on the thread to try to remain as independent as possible in my reasoning, so there may be ideas duplicated here.

Overall, I think this is an excellent package and provides useful functionality, which is well-documented.  Most of the notes below are optional suggestions based on my subjective experience of trying out the package.

The minimal changes I'd suggest incorporating are:

* adding the code of conduct
* updating the vignette which didn't render successfully for me locally

#### Running code examples in the README

Summary: looks great - a few suggestions below but feel free to treat them all as optional.

I tried to set things up to try the package out myself, using the instructions in the README.

I did feel a little uncomfortable entering in my SAS online password into the command line - I would have preferred to be given information about the config file and where it's stored to enter it there myself as that seems more typical of how other packages handle this kind of thing.

The password remains in the R terminal unobscured after being entered, which isn't great security practice.

When I went to configure it, the setup script asked "Which server is your account on?" with the following options:

```
1. United States 1
2. United States 2
3. Europe 2
4. Asia Pacific 1
5. Asia Pacific 2
```

Mine is "Europe 1" but I was unable to select this.  This didn't prevent it from running in the end though.

I also got this.

```
✔ Writing to C:\Users\thisi\OneDrive\DOCUME~1\VIRTUA~1\r-saspy\lib\site-packages\saspy/sascfg_personal.py.
✔ Opening sascfg_personal.py.
ℹ For more information about SASPy configuration see the sasquatch::configuration vignette or the
  SASPy documentation.
No java installation found. Enter the java path manually within sascfg_personal.py.
Error: vignette ‘configuration’ not found
> sasquatch::configuration
Error: 'configuration' is not an exported object from 'namespace:sasquatch
```

This confused me as, skimming, my instinct is to try to run `sasquatch::configuration` though this isn't the correct syntax for viewing a vignette.

Once I had done this and tried again with Java installed, it worked great!  I would have liked to have the SAS code example from the animated gif in the code example so I could copy and paste it.

I tried another example from the README

> sas_from_r(mtcars, "mtcars")
In sas_from_r(mtcars, "mtcars") :
  `x` rownames will not be transferred as a column.

It took me a moment to understand that `x` refers to the source table - this could be rephrased to be clearer, or a different example used which doesn't result in a warning message.

It'd also be handy to have it explained explicitly that this creates it as "mtcars" in the current SAS session - this wasn't clear to me and I had to check the docs to find this out.

> cars <- sas_to_r("cars", libref = "sashelp")

As there is already a dataset in R called `cars` I might be tempted to call this variable something else.

Overall, I found the instructions in the README clear enough to follow.  Although I had a few small technical hiccups, it worked and I was impressed by being able to run SAS code in a Quarto markdown document.

#### Running `devtools::check()`

```
❯ checking package subdirectories ... NOTE
  Problems with news in 'NEWS.md':
  No news entries found.

❯ checking R code for possible problems ... NOTE
  .rlang_check_is_string: no visible binding for global variable 'na_chr'
  .rlang_stop_unexpected_typeof: no visible global function definition
    for 'caller_env'
  check_logical: no visible global function definition for 'caller_arg'
  check_logical: no visible global function definition for 'caller_env'
  stop_input_type: no visible global function definition for 'caller_arg'
  stop_input_type: no visible global function definition for 'caller_env'
    caller_arg caller_env na_chr
```

Looking at the second of those, it looks like the notes were generated by the signature for `check_logical()`

https://github.com/ryanzomorrodi/sasquatch/blob/18cd7ca03a107be78700fdd3febf6908f28563c6/R/import-standalone-types-check.R#L508-L509

I might be tempted to write my own wrapper function or ask rlang to export their check functions, but this is just personal preference.

When I build the "configuration.rmd" vignette locally, for the first code block under "Executing SAS code", it just dumps out the HTML like so:

```
sas_run_string("PROC MEANS DATA = sashelp.cars;RUN;")
#> $x
#> $x$lst
#> [1] "<!DOCTYPE html>\n<html lang=\"en\" xml:lang=\"en\" xmlns=\"http://www.w3.org/1999/xhtml\">\n<head>\n<meta charset=\"utf-8\"/>\n<meta content=\"SAS 9.4\" name=\"generator\"/>\n<title>SAS Output</title>\n<style>\n/*<![CDATA[*/\n.body.c > table, .body.c > pre, .body.c div > table,\n.body.c div > pre, .body.c > table, .body.c > pre,\n.body.j > table, .body.j > pre, .body.j div > table,\n.body.j div > pre, .body.j > table, .body.j > pre,\n.body.c p.note, .body.c p.warning, .body.c p.error, .body.c p.fatal,\n.body.j p.note, .body.j p.warning, .body.j p.error, .body.j p.fatal,\n.body.c > table.layoutcontainer, .body.j > table.layoutcontainer { margin-left: auto; margin-right: auto }\n.layoutregion.l table, .layoutregion.l pre, .layoutregion.l p.note,\n.layoutregion.l p.warning, .layoutregion.l p.error, .layoutregion.l p.fatal { margin-left: 0 }\n.layoutregion.c table, .layoutregion.c pre, .layoutregion.c p.note,\n.layoutregion.c p.warning, .layoutregion.c p.error, .layoutregion.c p.fatal { margin-left: auto; margin-right: auto }\n.layoutregion.r table, .layoutregion.r pre, .layoutregion.r p.note,\n.layoutregion.r table, .layoutregion.r pre, .layoutregion.r p.note,\n.layoutregion.r p.warning, .layoutregion.r p.error, .layoutregion.r p.fatal { margin-right: 0 }\narticle, aside, details, figcaption, figure, footer, header, hgroup, nav, section { display: block }\nhtml{ font-size: 100% }\n.body { margin: 1em; font-size: 13px; line-height: 1.231 }\nsup { position: relative; vertical-align: baseline; bottom: 0.25em; font-size: 0.8em }\nsub { position: relative; vertical-align: baseline; top: 0.25em; font-size: 0.8em }\nul, ol { margin: 1em 0; padding: 0 0 0 40px }\ndd { margin: 0 0 0 40px }\nnav ul, nav ol { list-style: none; list-style-image: none; margin: 0; padding: 0 }\nimg { border: 0; vertical-align: middle }\nsvg:not(:root) { overflow: hidden }\nfigure { margin: 0 }\ntable { border-collapse: collapse;
...
```

It might be worth including this output as an image or other workaround instead of as runnable code.

In the configuration vignette, the "Access methods" contains a nested list - it's a little hard to skim and might be better as different headings instead.

#### The exported functions

I found the naming of the exported functions intuitive.  The exported functionality feels like enough to get things done without being overwhelming.  I think as the package evolves and users asks questions, there will be opportunity to expand some of the descriptions of functions to contain a little more information.

#### Contribution guidelines

These guidelines exist, but the code of conduct that they link to is missing.
