### 2025-05-05 #673 @tataphani
got this email

On Sun, May 4, 2025 at 3:08 PM Ryan Zomorrodi ***@***.***>
wrote:

> *ryanzomorrodi* created an issue (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673>
>
> Submitting Author Name: Ryan Zomorrodi
> Submitting Author Github Handle: @ryanzomorrodi
> <https://github.com/ryanzomorrodi>
> Other Package Authors Github handles: (comma separated, delete if none)
> Repository: https://github.com/ryanzomorrodi/sasquatch
> Version submitted: 0.0.0.9004
> Submission type: Standard
> Editor: @maurolepore <https://github.com/maurolepore>
> Reviewers: @thisisnic <https://github.com/thisisnic>, @tataphani
> <https://github.com/tataphani>
> Due date for @tataphani <https://github.com/tataphani>: 2025-05-25
>
> Archive: TBD
> Version accepted: TBD
> Language: en
> ------------------------------
>
>    - Paste the full DESCRIPTION file inside a code block below:
>
> Package: sasquatch
> Title: Use 'SAS', R, and 'quarto' Together
> Version: 0.0.0.9004
> ***@***.***:
>     person("Ryan", "Zomorrodi", , ***@***.***", role = c("aut", "cre", "cph"),
>            comment = c(ORCID = "0009-0003-6417-5985"))
> Description: Use R and 'SAS' within reproducible multilingual 'quarto'
>     documents. Run 'SAS' code blocks interactively, send data back and forth
>     between 'SAS' and R, and render 'SAS' output within quarto documents. 'SAS'
>     connections are established through a combination of 'SASPy' and
>     'reticulate'.
> License: MIT + file LICENSE
> Encoding: UTF-8
> Roxygen: list(markdown = TRUE)
> RoxygenNote: 7.3.2
> SystemRequirements:
>     python (>= 3.0.0),
>     SASPy,
>     java (>= 7) required for IOM access method
> Imports:
>     htmlwidgets,
>     knitr,
>     reticulate,
>     rstudioapi
> URL: https://ryanzomorrodi.github.io/sasquatch/, https://github.com/ryanzomorrodi/sasr
> Suggests:
>     rmarkdown,
>     testthat (>= 3.0.0)
> VignetteBuilder: knitr
> Config/testthat/edition: 3
> BugReports: https://github.com/ryanzomorrodi/sasr/issues
>
> Scope
>
>    -
>
>    Please indicate which category or categories from our package fit
>    policies
>    <https://devguide.ropensci.org/softwarereview_policies.html#package-categories>
>    this package falls under: (Please check an appropriate box below. If you
>    are unsure, we suggest you make a pre-submission inquiry.):
>    - data retrieval
>       - data extraction
>       - data munging
>       - data deposition
>       - data validation and testing
>       - workflow automation
>       - version control
>       - citation management and bibliometrics
>       - scientific software wrappers
>       - field and lab reproducibility tools
>       - database software bindings
>       - geospatial data
>       - text analysis
>    -
>
>    Explain how and why the package falls under these categories (briefly,
>    1-2 sentences):
>
> sasquatch allows users to create remote SAS sessions, work with them
> interactively, pass tables between R and SAS, and render quarto documents.
>
>    - Who is the target audience and what are scientific applications of
>    this package?
>
> The target audience would be anyone looking to combine SAS in R within a
> single workflow. Using quarto, analysts can have one single reproducable
> report that combines SAS and R (and python potentially).
>
>    - Are there other R packages that accomplish the same thing? If so,
>    how does yours differ or meet our criteria for best-in-category
>    <https://ropensci.github.io/dev_guide/policies.html#overlap>?
>
> saquatch works similarly to packages like sasr
> <https://sassoftware.github.io/saspy/> or configSAS
> <https://github.com/baselr/configSAS>. sasr, while similar to sasquatch,
> does not include interactive SAS functionality or a knitr engine. On the
> other hand, configSAS includes a knitr engine, but no interactive SAS
> functionality. configSAS knitr output also does not include syntax
> highlighting and nested SAS output interferes with the styles of the rest
> of the document.
>
>    - (If applicable) Does your package comply with our guidance around *Ethics,
>    Data Privacy and Human Subjects Research*
>    <https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research>
>    ?
>
> N/A
>
>    -
>
>    If you made a pre-submission inquiry, please paste the link to the
>    corresponding issue, forum post, or other discussion, or @tag
>    <https://github.com/tag> the editor you contacted.
>    -
>
>    Explain reasons for any pkgcheck items
>    <https://docs.ropensci.org/pkgcheck/> which your package is unable to
>    pass.
>
> Locally, pkgcheck is able to pass all tests except coverage. I do not
> currently have any coverage over the knitr engine because I am struggling
> to think about how I can test that effectively. My assumption that the
> automated pkgcheck testing will fail due to the need for python, python
> package SASPy, java, and a SAS License. (I have set up a CI that has all
> of these set up within the sasquatch repo).
> Technical checks
>
> Confirm each of the following by checking the box.
>
>    - I have read the rOpenSci packaging guide
>    <https://devguide.ropensci.org/building.html>.
>    - I have read the author guide
>    <https://devdevguide.netlify.app/authors-guide.html> and I expect to
>    maintain this package for at least 2 years or to find a replacement.
>
> This package:
>
>    - does not violate the Terms of Service of any service it interacts
>    with.
>    - has a CRAN and OSI accepted license.
>    - contains a README with instructions for installing the development
>    version <https://ropensci.github.io/dev_guide/building.html#readme>.
>    - includes documentation with examples for all functions, created with
>    roxygen2
>    <https://ropensci.github.io/dev_guide/building.html#documentation>.
>    - contains a vignette with examples of its essential functions and
>    uses. (this would need to be rendered locally)
>    - has a test suite
>    <https://ropensci.github.io/dev_guide/building.html#testing>.
>    - has continuous integration
>    <https://ropensci.github.io/dev_guide/ci.html>, including reporting of
>    test coverage.
>
> Publication options
>
>    -
>
>    Do you intend for this package to go on CRAN?
>    -
>
>    Do you intend for this package to go on Bioconductor?
>    -
>
>    Do you wish to submit an Applications Article about your package to Methods
>    in Ecology and Evolution
>    <http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/>?
>    If so:
>
> MEE Options
>
>    - The package is novel and will be of interest to the broad readership
>    of the journal.
>    - The manuscript describing the package is no longer than 3000 words.
>    - You intend to archive the code for the package in a long-term
>    repository which meets the requirements of the journal (see MEE's
>    Policy on Publishing Code
>    <http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html>
>    )
>    - (*Scope: Do consider MEE's Aims and Scope
>    <http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html>
>    for your manuscript. We make no guarantee that your manuscript will be
>    within MEE scope.*)
>    - (*Although not required, we strongly recommend having a full
>    manuscript prepared when you submit here.*)
>    - (*Please do not submit your package separately to Methods in Ecology
>    and Evolution*)
>
> Code of conduct
>
>    - I agree to abide by rOpenSci's Code of Conduct
>    <https://ropensci.org/code-of-conduct/> during the review process and
>    in maintaining my package should it be accepted.
>
