### 2025-05-06 #673 @maurolepore
@tataphani can you summarize what the email says if you believe is relevant?
