### 2025-05-29 #673 @tataphani
Yes, I am working on it. Just got very busy with the day Job , very sorry
for the delay . Just curious about the steps that I need to check other
things can I contact you on slack and have an extension of the date

On Wed, May 28, 2025 at 7:34 PM Mauro Lepore ***@***.***>
wrote:

> *maurolepore* left a comment (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673#issuecomment-2917912784>
>
> @tataphani <https://github.com/tataphani> the bot gently reminded us
> about your review. Could you please give us an estimate of when it would be
> ready?
>
