### 2025-06-15 ryanzomorrodi/sasquatch#8 @tataphani
installation problem
I have already installed Python and had my access to SAS online edtion while trying to install this error see the log and it does not suggest any other dependencies needed can you please help

> pak::pkg_install("ryanzomorrodi/sasquatch")
! Using bundled GitHub PAT. Please add your own PAT using `gitcreds::gitcreds_set()`.
✔ Updated metadata database: 5.43 MB in 9 files.
✔ Updating metadata database ... done

→ Will install 25 packages.
→ Will download 24 CRAN packages (18.43 MB).
→ Will download 1 package with unknown size.
+ base64enc     0.1-3      [dl] (32.65 kB)
+ bslib         0.9.0      [dl] (5.92 MB)
+ cachem        1.1.0      [dl] (72.55 kB)
+ digest        0.6.37     [dl] (221.79 kB)
+ evaluate      1.0.3      [dl] (102.34 kB)
+ fastmap       1.2.0      [dl] (133.27 kB)
+ fontawesome   0.5.3      [dl] (1.40 MB)
+ highr         0.11       [dl] (44.03 kB)
+ htmlwidgets   1.6.4      [dl] (812.12 kB)
+ jquerylib     0.1.4      [dl] (526.00 kB)
+ jsonlite      2.0.0      [dl] (1.11 MB)
+ knitr         1.50       [dl] (1.15 MB)
+ lifecycle     1.0.4      [dl] (139.76 kB)
+ memoise       2.0.1      [dl] (50.00 kB)
+ mime          0.13       [dl] (52.43 kB)
+ R6            2.6.1      [dl] (87.95 kB)
+ rappdirs      0.3.3      [dl] (51.42 kB)
+ Rcpp          1.0.14     [dl] (2.89 MB)
+ rmarkdown     2.29       [dl] (2.69 MB)
+ rprojroot     2.0.4      [dl] (113.50 kB)
+ rstudioapi    0.17.1     [dl] (334.21 kB)
+ sasquatch     0.0.0.9042 [bld][cmp][dl] (GitHub: 18cd7ca)
+ tinytex       0.56       [dl] (145.07 kB)
+ withr         3.0.2      [dl] (228.26 kB)
+ yaml          2.3.10     [dl] (118.41 kB)
ℹ Getting 24 pkgs (18.43 MB) and 1 pkg with unknown size
✔ Got R6 2.6.1 (i386+x86_64-w64-mingw32) (87.88 kB)
✔ Got base64enc 0.1-3 (x86_64-w64-mingw32) (32.65 kB)
✔ Got cachem 1.1.0 (x86_64-w64-mingw32) (72.77 kB)
✔ Got digest 0.6.37 (x86_64-w64-mingw32) (222.07 kB)
✔ Got evaluate 1.0.3 (i386+x86_64-w64-mingw32) (102.43 kB)
✔ Got fastmap 1.2.0 (x86_64-w64-mingw32) (133.49 kB)
✔ Got highr 0.11 (i386+x86_64-w64-mingw32) (44.15 kB)
✔ Got tinytex 0.56 (i386+x86_64-w64-mingw32) (145.04 kB)
✔ Got rprojroot 2.0.4 (i386+x86_64-w64-mingw32) (114.01 kB)
✔ Got rstudioapi 0.17.1 (i386+x86_64-w64-mingw32) (336.93 kB)
✔ Got Rcpp 1.0.14 (x86_64-w64-mingw32) (2.89 MB)
✔ Got bslib 0.9.0 (i386+x86_64-w64-mingw32) (5.92 MB)
✔ Got jquerylib 0.1.4 (i386+x86_64-w64-mingw32) (526.07 kB)
✔ Got mime 0.13 (x86_64-w64-mingw32) (52.43 kB)
✔ Got rappdirs 0.3.3 (x86_64-w64-mingw32) (51.75 kB)
✔ Got fontawesome 0.5.3 (i386+x86_64-w64-mingw32) (1.40 MB)
✔ Got knitr 1.50 (i386+x86_64-w64-mingw32) (1.15 MB)
✔ Got lifecycle 1.0.4 (i386+x86_64-w64-mingw32) (140.13 kB)
✔ Got memoise 2.0.1 (i386+x86_64-w64-mingw32) (50.70 kB)
✔ Got htmlwidgets 1.6.4 (i386+x86_64-w64-mingw32) (812.97 kB)
✔ Got rmarkdown 2.29 (i386+x86_64-w64-mingw32) (2.69 MB)
✔ Got yaml 2.3.10 (x86_64-w64-mingw32) (118.65 kB)
✔ Got withr 3.0.2 (i386+x86_64-w64-mingw32) (228.35 kB)
✔ Got jsonlite 2.0.0 (x86_64-w64-mingw32) (1.11 MB)
✔ Got sasquatch 0.0.0.9042 (source) (1.67 MB)
✔ Downloaded 25 packages (20.10 MB) in 2.7s
✔ Installed R6 2.6.1  (653ms)
✔ Installed base64enc 0.1-3  (818ms)
✔ Installed cachem 1.1.0  (1.2s)
✔ Installed evaluate 1.0.3  (1.6s)
✔ Installed fastmap 1.2.0  (1.9s)
✔ Installed digest 0.6.37  (2.2s)
✔ Installed fontawesome 0.5.3  (1.9s)
✔ Installed highr 0.11  (2s)
✔ Installed htmlwidgets 1.6.4  (1.8s)
✔ Installed jquerylib 0.1.4  (1.8s)
✔ Installed jsonlite 2.0.0  (2.1s)
✔ Installed lifecycle 1.0.4  (1.9s)
✔ Installed memoise 2.0.1  (2s)
✔ Installed knitr 1.50  (2.9s)
✔ Installed Rcpp 1.0.14  (5.6s)
✔ Installed mime 0.13  (2.5s)
✔ Installed rappdirs 0.3.3  (2.4s)
✔ Installed bslib 0.9.0  (6.5s)
✔ Installed rprojroot 2.0.4  (2s)
✔ Installed rstudioapi 0.17.1  (2.1s)
✔ Installed tinytex 0.56  (2s)
✔ Installed rmarkdown 2.29  (3.2s)
✔ Installed withr 3.0.2  (1.9s)
✔ Installed yaml 2.3.10  (1.7s)
ℹ Packaging sasquatch 0.0.0.9042
✔ Packaged sasquatch 0.0.0.9042 (1.7s)
ℹ Building sasquatch 0.0.0.9042
Error:
! error in pak subprocess
Caused by error:
! Could not find tools necessary to compile a package
Call `pkgbuild::check_build_tools(debug = TRUE)` to diagnose the problem.
Type .Last.error to see the more details.
