### 2025-06-16 ryanzomorrodi/sasquatch#8 @ryanzomorrodi
I added `sasquatch` to r-universe last week so I think you should just be able to install using:

```r
install.packages('sasquatch', repos = c('https://ryanzomorrodi.r-universe.dev', 'https://cloud.r-project.org'))
```

Typically, in order to install any R package from source on Windows you will have to install Rtools. Since you are using R 4.3.3 you would have to install [Rtools43](https://cran.r-project.org/bin/windows/Rtools/rtools43/rtools.html). But the r-universe version should work for you.
