### 2025-06-16 ryanzomorrodi/sasquatch#8 @ryanzomorrodi
I'm not sure. Since you installed Rtools, hopefully you should be able to run the original:

```r
pak::pkg_install("ryanzomorrodi/sasquatch")
```

If that doesn't work then:

> Call pkgbuild::check_build_tools(debug = TRUE) to diagnose the problem.
> Type .Last.error to see the more details.

And send me the output of these messages
