### 2025-06-16 ryanzomorrodi/sasquatch#8 @tataphani
> pak::pkg_install("ryanzomorrodi/sasquatch")! Using bundled GitHub PAT. Please add your own PAT using `gitcreds::gitcreds_set()`.✔ Loading metadata database ... done

     ℹ No downloads are needed✔ 1 pkg + 37 deps: kept 31 [11s]>
library(sasquatch)> reticulate::install_python()Error in
pyenv_bootstrap_windows() :
  Please install git and ensure it is on your PATH>
library(reticulate)> sasquatch::install_saspy()Using Python:
C:/Users/tatap/AppData/Local/Programs/Python/Python313/python.exe
Creating virtual environment "r-saspy" ... +
"C:/Users/tatap/AppData/Local/Programs/Python/Python313/python.exe" -m
venv "C:/Users/tatap/OneDrive/Documents/.virtualenvs/r-saspy"Done!
Installing packages: pip, wheel, setuptools+
"C:/Users/tatap/OneDrive/Documents/.virtualenvs/r-saspy/Scripts/python.exe"
-m pip install --upgrade pip wheel setuptoolsRequirement already
satisfied: pip in
c:\users\tatap\onedrive\docume~1\virtua~1\r-saspy\lib\site-packages
(25.0.1)
Collecting pip
  Downloading pip-25.1.1-py3-none-any.whl.metadata (3.6 kB)
Collecting wheel
  Downloading wheel-0.45.1-py3-none-any.whl.metadata (2.3 kB)
Collecting setuptools
  Downloading setuptools-80.9.0-py3-none-any.whl.metadata (6.6 kB)
Downloading pip-25.1.1-py3-none-any.whl (1.8 MB)
   ---------------------------------------- 1.8/1.8 MB 11.4 MB/s eta 0:00:00
Downloading wheel-0.45.1-py3-none-any.whl (72 kB)
Downloading setuptools-80.9.0-py3-none-any.whl (1.2 MB)
   ---------------------------------------- 1.2/1.2 MB 16.3 MB/s eta 0:00:00
Installing collected packages: wheel, setuptools, pip
  Attempting uninstall: pip
    Found existing installation: pip 25.0.1
    Uninstalling pip-25.0.1:
      Successfully uninstalled pip-25.0.1
Successfully installed pip-25.1.1 setuptools-80.9.0 wheel-0.45.1
Virtual environment 'r-saspy' successfully created.
Using virtual environment "r-saspy" ...+
"C:/Users/tatap/OneDrive/Documents/.virtualenvs/r-saspy/Scripts/python.exe"
-m pip install --upgrade --no-user wheel saspy pandasRequirement
already satisfied: wheel in
c:\users\tatap\onedrive\docume~1\virtua~1\r-saspy\lib\site-packages
(0.45.1)
Collecting saspy
  Downloading saspy-5.103.0-py3-none-any.whl.metadata (3.9 kB)
Collecting pandas
  Downloading pandas-2.3.0-cp313-cp313-win_amd64.whl.metadata (19 kB)
Collecting numpy>=1.26.0 (from pandas)
  Downloading numpy-2.3.0-cp313-cp313-win_amd64.whl.metadata (60 kB)
Collecting python-dateutil>=2.8.2 (from pandas)
  Using cached python_dateutil-2.9.0.post0-py2.py3-none-any.whl.metadata
(8.4 kB)
Collecting pytz>=2020.1 (from pandas)
  Downloading pytz-2025.2-py2.py3-none-any.whl.metadata (22 kB)
Collecting tzdata>=2022.7 (from pandas)
  Downloading tzdata-2025.2-py2.py3-none-any.whl.metadata (1.4 kB)
Collecting six>=1.5 (from python-dateutil>=2.8.2->pandas)
  Using cached six-1.17.0-py2.py3-none-any.whl.metadata (1.7 kB)
Downloading saspy-5.103.0-py3-none-any.whl (10.0 MB)
   ---------------------------------------- 10.0/10.0 MB 21.5 MB/s eta 0:00:00
Downloading pandas-2.3.0-cp313-cp313-win_amd64.whl (11.0 MB)
   ---------------------------------------- 11.0/11.0 MB 54.1 MB/s eta 0:00:00
Downloading numpy-2.3.0-cp313-cp313-win_amd64.whl (12.7 MB)
   ---------------------------------------- 12.7/12.7 MB 57.6 MB/s eta 0:00:00
Using cached python_dateutil-2.9.0.post0-py2.py3-none-any.whl (229 kB)
Downloading pytz-2025.2-py2.py3-none-any.whl (509 kB)
Using cached six-1.17.0-py2.py3-none-any.whl (11 kB)
Downloading tzdata-2025.2-py2.py3-none-any.whl (347 kB)
Installing collected packages: saspy, pytz, tzdata, six, numpy,
python-dateutil, pandas

Successfully installed numpy-2.3.0 pandas-2.3.0
python-dateutil-2.9.0.post0 pytz-2025.2 saspy-5.103.0 six-1.17.0
tzdata-2025.2Warning messages:
1: In file.info(x, extra_cols = FALSE) :
  cannot open file
'C:/Users/tatap/AppData/Local/Microsoft/WindowsApps/python3.exe': The
file cannot be accessed by the system
2: In file.info(x, extra_cols = FALSE) :
  cannot open file
'C:/Users/tatap/AppData/Local/Microsoft/WindowsApps/python3.exe': The
file cannot be accessed by the system
Restarting R session...
> sasquatch::configure_saspy(template = "oda")Enter your ODA username: KRISHNA NEMANIEnter your ODA password: Jaigurudatta_1627✔ Writing to C:\Users\tatap/_authinfo.
Which server is your account on?

1. United States 1
2. United States 2
3. Europe 2
4. Asia Pacific 1
5. Asia Pacific 2Selection: 1✔ Writing to /sascfg_personal.py.Error in
file(file, ifelse(append, "a", "w")) :
  cannot open the connection
In addition: Warning messages:
1: No java installation found. Enter the java path manually within
sascfg_personal.py.
2: In file(file, ifelse(append, "a", "w")) :
  cannot open file '/sascfg_personal.py': Permission denied>
install.packages("saspy")Installing package into
‘C:/Users/tatap/AppData/Local/R/win-library/4.3’
(as ‘lib’ is unspecified)Warning in install.packages :
  package ‘saspy’ is not available for this version of R

A version of this package for your version of R might be available elsewhere,
see the ideas athttps://cran.r-project.org/doc/manuals/r-patched/R-admin.html#Installing-packages

>

My Python is installed at this location but it is still complaining
1.8 1.8.0_451 http://java.sun.com/products/autodl/j2se *C:\Program
Files\Java\jre1.8.0_451\bin\javaw.exe* true x86_64

[image: image.png]
