### 2025-06-17 ryanzomorrodi/sasquatch#8 @ryanzomorrodi
Hi @tataphani . Definitely try to be careful with whatever you send to me. Even if you send a reply via email, all your replies are posted on the public github issue. It seems you included your SAS ODA password. I would definitely recommend resetting your password, and hopefully you do not reuse passwords, but if you do, I would recommend resetting those as well.

I have deleted your last reply, but because of the nature of the internet I cannot guarantee that no one else saw it in the meantime or that it has not been saved somewhere else.
