### 2025-06-17 ryanzomorrodi/sasquatch#8 @ryanzomorrodi
As to your issue, there are a couple things at play here. So

1.  reticulate is not detecting your install of `saspy`.
2. Your java installation path is not found

Let's do some troubleshooting of the first problem, first

It seems that `saspy` was installed correctly. But I wanna just double check:

Run
```r
reticulate::py_discover_config("saspy", "r-saspy")
```

If you get
``` txt
Error:
! Virtual environment 'r-saspy' does not exist.
```
Then, `sasquatch::install_saspy()` either did not create the virtual environment correctly or `reticulate` cannot find the virtual env. Either way let me know.

You should get an output like this:
```txt
python:         /home/user/.virtualenvs/r-saspy/bin/python
libpython:      /usr/lib/python3.12/config-3.12-x86_64-linux-gnu/libpython3.12.so
pythonhome:     /home/user/.virtualenvs/r-saspy:/home/user/.virtualenvs/r-saspy
version:        3.12.3 (main, Feb  4 2025, 14:48:35) [GCC 13.3.0]
numpy:          /home/user/.virtualenvs/r-saspy/lib/python3.12/site-packages/numpy
numpy_version:  2.2.4
saspy:          /home/user/.virtualenvs/r-saspy/lib/python3.12/site-packages/saspy
```

If you see
```txt
saspy:          [NOT FOUND]
```

Then the virtual environment was created, but saspy is not in it.

Ideally, you should see something like this:

```txt
saspy:          C:/Users/tatap/OneDrive/Documents/.virtualenvs/r-saspy/lib/site-packages/saspy
```
