### 2025-06-22 #673 @maurolepore
For the record, linking an installation issue that was solved in sasquatch itself.

* https://github.com/ryanzomorrodi/sasquatch/issues/8

The relevant guideline is this:

> You may also submit issues or pull requests directly to the package repo if you choose, but if you do, please comment about them and link to them in the software-review repo comment thread so we have a centralized record and text of your review.
https://devdevguide.netlify.app/softwarereview_reviewer#submitreview

@ryanzomorrodi thanks so much for helping with that.

@tataphani thanks so much for sharing your experience.
