### 2025-06-28 c27012b
Add warning for no-SASPy installation
