### 2025-06-28 #673 @maurolepore
@tataphani how do you feel about the review process? Are you still able to review or did other priorities get in the way?
