### 2025-06-29 #673 @tataphani
Hello Mauro ,

I am little confused , I did ask for the help with  sas_connect_id   part
and mailed same thing 2 weeks ago but you replied back   it was under issue
# 8 correspondence ,
Issue #
I am struck at this point at sas_connect id , any idea where can I find
this id

Status is that I can't install the package with out sas_connect_id

Thanks
Phani

On Sat, Jun 28, 2025 at 6:47 PM Mauro Lepore ***@***.***>
wrote:

> *maurolepore* left a comment (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673#issuecomment-3016154621>
>
> @tataphani <https://github.com/tataphani> how do you feel about the
> review process? Are you still able to review or did other priorities get in
> the way?
>
