+++ R/saspy-configure.R
@@ -59,43 +59,15 @@ configure_saspy <- function(
   template = c("none", "oda"),
   overwrite = FALSE
 ) {
-  if (template[1] == "none") {
+  template <- rlang::arg_match(template, c("none", "oda"))
+
+  saspy_path <- get_saspy_path()
+
+  if (template == "none") {
     configs <- list(config_name = list())
   } else if (template == "oda") {
-    oda_username <- readline("Enter your ODA username: ")
-    oda_password <- readline("Enter your ODA password: ")
-    write_authinfo("oda", oda_username, oda_password, overwrite)
-
-    java_path <- Sys.which("java")
-    if (identical(unname(java_path), "")) {
-      cli::cli_warn(
-        "No java installation found. Enter the java path manually within sascfg_personal.py."
-      )
-    }
-
-    oda_servers <- list(
-      "United States 1" = list(
-        'odaws01-usw2.oda.sas.com',
-        'odaws02-usw2.oda.sas.com',
-        'odaws03-usw2.oda.sas.com',
-        'odaws04-usw2.oda.sas.com'
-      ),
-      "United States 2" = list(
-        'odaws01-usw2-2.oda.sas.com',
-        'odaws02-usw2-2.oda.sas.com'
-      ),
-      "Europe 1" = list('odaws01-euw1.oda.sas.com', 'odaws02-euw1.oda.sas.com'),
-      "Asia Pacific 1" = list(
-        'odaws01-apse1.oda.sas.com',
-        'odaws02-apse1.oda.sas.com'
-      ),
-      "Asia Pacific 2" = list(
-        'odaws01-apse1-2.oda.sas.com',
-        'odaws02-apse1-2.oda.sas.com'
-      )
-    )
-    server_num = menu(names(oda_servers), "Which server is your account on?")
-    iom_host = oda_servers[[server_num]]
+    java_path <- check_java_installation(saspy_path)
+    iom_host <- ask_oda_server()
 
     configs <- list(
       oda = list(
@@ -108,12 +80,16 @@ configure_saspy <- function(
     )
   }
 
-  sascfg_personal_path <- write_sascfg_personal(configs, overwrite)
+  write_sascfg_personal(configs, saspy_path, overwrite)
 
-  if (rstudioapi::hasFun("navigateToFile")) {
-    cli::cli_alert_success("Opening sascfg_personal.py.")
-    rstudioapi::navigateToFile(sascfg_personal_path)
+  if (template == "oda") {
+    write_authinfo_template(overwrite)
+    sas_oda_jar_dir <- paste0(saspy_path, "/java/iomclient/")
+    cli::cli_inform(c(
+      "i" = "Download and extract the SAS ODA zip encryption jars ({.url https://support.sas.com/downloads/package.htm?pid=2494}) into {.file {sas_oda_jar_dir}}."
+    ))
   }
+
   cli::cli_inform(c(
     "i" = "For more information about {.pkg SASPy} configuration see the {.vignette sasquatch::configuration} vignette or the {.href [SASPy documentation](https://sassoftware.github.io/saspy/configuration.html)}."
   ))
@@ -121,18 +97,64 @@ configure_saspy <- function(
   invisible()
 }
 
-get_home_dir <- function() {
-  home_dir <- Sys.getenv("HOME")
-  if (.Platform$OS.type == "windows") {
-    home_dir <- regmatches(home_dir, regexpr("(.*?[/|\\\\]){3}", home_dir))
+get_saspy_path <- function() {
+  python <- reticulate::py_discover_config("saspy", "r-saspy")
+  saspy_path <- python$required_module_path
[…]
