### 2025-06-30 #673 @maurolepore
@tataphani thanks for your quick response.

To be most helpful please ensure to share any issues in this main thread (see also https://github.com/ropensci/software-review/issues/673#issuecomment-2993959313). (Another tiny details that might help is to communicate using the GitHub UI rather than email. Emails tend to duplicate information which might obscure the latest most relevant details.)

At least that can help eliminate any confusion about where the conversation is held, and leave us only with the more interesting aspects of the challenge.
