### 2025-07-09 #673 @ryanzomorrodi
@maurolepore Would it be okay if I started merging commits that addressed thisisnic's suggestions? I know you earlier mentioned you didn't want the review to feel like a moving target, but I have made some improvements on a different branch that improve the setup experience.

I'm not quite sure at what step @tataphani is caught up on, but I did realize when I did an install on a windows machine that the connecting to ODA sometimes requires a few additional java files, so I added a message for that, and removed the username password questions.

May be this may help tataphani or could be more clear in the documentation, but the steps for setup of ODA are as follows.
1. Install sasquatch (If on windows, you will need to download RTools first to install from source)
2. Install python `reticulate::install_python()`  (if not already installed)
3. Install java (if not already installed). Adding java to path also helps because then `sasquatch` can find it's path easier.
4.  (Optional) `reticulate::py_available()` (might be good just to see if your python install went well and is able to be found by reticulate)
5. `sasquatch::install_saspy()`
6. `sasquatch::config_saspy("oda")`
   This process may be a little confusing, but ideally you should get a `sascfg_personal.py` file inside of your `saspy` installation that looks like:
   ```python
   SAS_config_names=['oda']
   oda = {
     'java' : 'your_java_path',
     # Uncomment the one for your region
     #US Home Region 1
     # 'iomhost' : ['odaws01-usw2.oda.sas.com','odaws02-usw2.oda.sas.com','odaws03-usw2.oda.sas.com','odaws04-usw2.oda.sas.com'],
     #US Home Region 2
     #'iomhost' : ['odaws01-usw2-2.oda.sas.com','odaws02-usw2-2.oda.sas.com'],
     #European Home Region 1
     #'iomhost' : ['odaws01-euw1.oda.sas.com','odaws02-euw1.oda.sas.com'],
     #Asia Pacific Home Region 1
     #'iomhost' : ['odaws01-apse1.oda.sas.com','odaws02-apse1.oda.sas.com'],
     #Asia Pacific Home Region 2
     #'iomhost' : ['odaws01-apse1-2.oda.sas.com','odaws02-apse1-2.oda.sas.com'],
     'iomport' : 8591,
     'authkey' : 'oda',
     'encoding' : 'utf-8'
   }
   ```
   and a `.authfile` (`_authfile`) in your system's home directory that contains your username and password like this:
   ```txt
   oda user {your username} password {your password}
   ```
   The last final complication is that you may have to download these [java encryption jars](https://support.sas.com/downloads/package.htm?pid=2494) and install them into the your saspy installation's `java/iomclient` folder.

As I write this I am thinking maybe there is some documentation I can improve, but please let me know if this helps. The changes I made more recently try to make this more clear when you run `configure_saspy()`, but let me know if there is still confusion.
