### 2025-07-09 #673 @ryanzomorrodi
Also it would be nice if we could get a new timeline for tataphani 's review. Since it is the summer, this is the best time for me to respond quickly to any concerns and I would like to use the time to try to push for CRAN submission, but I have been waiting for the review completion.
