### 2025-07-09 #673 @maurolepore
@ryanzomorrodi, thanks so much for your patience.

Considering the delay, your request seems very reasonable. Let me check with the editorial board. I was thinking of discussing this submission with the group and this is a good nudge in that direction.

I'll come back with an answer by the end of this Thursday.

In the meantime, @tataphani please let us know if you already started with the review, and when you estimate you'd be ready to submit it.

If not then merging commits would not disrupt his work.
