### 2025-07-09 #673 @tataphani
Hello Ropensci/Software-Review ,

I will restart installation process , I am not sure about the ID it is
requesting I am planning to use my Kids laptop install all the  required
packages including New version R and Rstudio , let's see if that helps

On Tue, Jul 8, 2025 at 8:35 PM Mauro Lepore ***@***.***>
wrote:

> *maurolepore* left a comment (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673#issuecomment-3050753218>
>
> @ryanzomorrodi <https://github.com/ryanzomorrodi>, thanks so much for
> your patience.
>
> Considering the delay, your request seems very reasonable. Let me check
> with the editorial board. I was thinking of discussing this submission with
> the group and this is a good nudge in that direction.
>
> I'll come back with an answer by the end of this Thursday.
>
> In the meantime, @tataphani <https://github.com/tataphani> please let us
> know if you already started with the review, and when you estimate you'd be
> ready to submit it.
>
> If not then merging commits would not disrupt his work.
>
