### 2025-07-11 b05862b
no more CI because jars are required for auth
