### 2025-07-11 #673 @ryanzomorrodi
I have merged the new changes.

It seems like the encryption jars I mentioned above are now required. Unfortunately, the means I had to get rid of the check and coverage CIs because to downloading the jars requires you to log in and accept the terms and conditions of use.
