### 2025-07-18 #673 @ryanzomorrodi
I would be cool with getting maybe a third opinion. The more eyes the better! Thank you for all your help @maurolepore
