### 2025-07-23 #673 @ryanzomorrodi
I am okay with moving to the next stages. I appreciate all your help Mauro and I'm sorry to @tataphani that we weren't able to help you get `sasquatch` set up. Thank you for your time and efforts.

The latest changes are as follows (this is also in the NEWS.md):
- `configure_saspy()` no longer asks that users type in their SAS ODA username and password within the terminal. See [secrets vignette](https://cran.r-project.org/web/packages/httr/vignettes/secrets.html).
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):

  > I did feel a little uncomfortable entering in my SAS online password into the command line - I would have preferred to be given information about the config file and where it's stored to enter it there myself as that seems more typical of how other packages handle this kind of thing.
  >
  > The password remains in the R terminal unobscured after being entered, which isn't great security practice.

- Dataset used to show `sas_to_r()` in `README.md` changed to `warpbreaks` and the variable name of the version of `"sashelp.cars"` was renamed to `sas_cars`.
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):

  > I tried another example from the README
  >
  > sas_from_r(mtcars, "mtcars")
  > Warning message:
  > In sas_from_r(mtcars, "mtcars") :
  > x rownames will not be transferred as a column.
  >
  > It took me a moment to understand that x refers to the source table - this could be rephrased to be clearer, or a different example used which doesn't result in a warning message.
  > It'd also be handy to have it explained explicitly that this creates it as "mtcars" in the current SAS session - this wasn't clear to me and I had to check the docs to find this out.
  >
  > cars <- sas_to_r("cars", libref = "sashelp")
  >
  > As there is already a dataset in R called cars I might be tempted to call this variable something else.

- Fixed typo in `configure_saspy()`. Changed Europe 2 to Europe 1.
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):

  > When I went to configure it, the setup script asked "Which server is your account on?" with the following options:
  >
  > 1. United States 1
  > 2. United States 2
  > 3. Europe 2
  > 4. Asia Pacific 1
  > 5. Asia Pacific 2
  >
  > Mine is "Europe 1" but I was unable to select this. This didn't prevent it from running in the end though.

- Specified `rlang` imports.
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):

  > Running devtools::check()
  >
  > \u276f checking package subdirectories ... NOTE
  >   Problems with news in 'NEWS.md':
  >   No news entries found.
  >
  > \u276f checking R code for possible problems ... NOTE
  >   .rlang_check_is_string: no visible binding for global variable 'na_chr'
  >   .rlang_stop_unexpected_typeof: no visible global function definition
  >     for 'caller_env'
  >   check_logical: no visible global function definition for 'caller_arg'
  >   check_logical: no visible global function definition for 'caller_env'
  >   stop_input_type: no visible global function definition for 'caller_arg'
  >   stop_input_type: no visible global function definition for 'caller_env'
  >   Undefined global functions or variables:
  >     caller_arg caller_env na_chr
  >
  > Looking at the second of those, it looks like the notes were generated by the signature for check_logical()
  >
  > https://github.com/ryanzomorrodi/sasquatch/blob/18cd7ca03a107be78700fdd3febf6908f28563c6/R/import-standalone-types-check.R#L508-L509
  >
  > I might be tempted to write my own wrapper function or ask rlang to export their check functions, but this is just personal preference.
- Added a Code of Conduct.
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):

  > Contribution guidelines
  >
  > These guidelines exist, but the code of conduct that they link to is missing.

Other notes:
- Kept `sasquatch::configuration` because it is the way that `cli` specifies vignettes. If you click on it within RStudio or Positron it should take you directly to the vignette.
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):
  > This confused me as, skimming, my instinct is to try to run sasquatch::configuration though this isn't the correct syntax for viewing a vignette.
- Could not reproduce this
  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):
  > When I build the "configuration.rmd" vignette locally, for the first code block under "Executing SAS code", it just dumps out the HTML like so:
  >
  > sas_run_string("PROC MEANS DATA = sashelp.cars;RUN;")
  > #> $x
  > #> $x$lst
  > #> [1] "<!DOCTYPE html>\n<html lang=\"en\" xml:lang=\"en\" xmlns=\"http://www.w3.org/1999/xhtml\">\n<head>\n<meta charset=\"utf-8\"/>\n<meta content=\"SAS 9.4\" name=\"generator\"/>\n<title>SAS Output</title>\n<style>\n/*<![CDATA[*/\n.body.c > table, .body.c > pre, .body.c div > table,\n.body.c div > pre, .body.c > table, .body.c > pre,\n.body.j > table, .body.j > pre, .body.j div > table,\n.body.j div > pre, .body.j > table, .body.j > pre,\n.body.c p.note, .body.c p.warning, .body.c p.error, .body.c p.fatal,\n.body.j p.note, .body.j p.warning, .body.j p.error, .body.j p.fatal,\n.body.c > table.layoutcontainer, .body.j > table.layoutcontainer { margin-left: auto; margin-right: auto }\n.layoutregion.l table, .layoutregion.l pre, .layoutregion.l p.note,\n.layoutregion.l p.warning, .layoutregion.l p.error, .layoutregion.l p.fatal { margin-left: 0 }\n.layoutregion.c table, .layoutregion.c pre, .layoutregion.c p.note,\n.layoutregion.c p.warning, .layoutregion.c p.error, .layoutregion.c p.fatal { margin-left: auto; margin-right: auto }\n.layoutregion.r table, .layoutregion.r pre, .layoutregion.r p.note,\n.layoutregion.r table, .layoutregion.r pre, .layoutregion.r p.note,\n.layoutregion.r p.warning, .layoutregion.r p.error, .layoutregion.r p.fatal { margin-right: 0 }\narticle, aside, details, figcaption, figure, footer, header, hgroup, nav, section { display: block }\nhtml{ font-size: 100% }\n.body { margin: 1em; font-size: 13px; line-height: 1.231 }\nsup { position: relative; vertical-align: baseline; bottom: 0.25em; font-size: 0.8em }\nsub { position: relative; vertical-align: baseline; top: 0.25em; font-size: 0.8em }\nul, ol { margin: 1em 0; padding: 0 0 0 40px }\ndd { margin: 0 0 0 40px }\nnav ul, nav ol { list-style: none; list-style-image: none; margin: 0; padding: 0 }\nimg { border: 0; vertical-align: middle }\nsvg:not(:root) { overflow: hidden }\nfigure { margin: 0 }\ntable { border-collapse: collapse;
...
  >
  > It might be worth including this output as an image or other workaround instead of as runnable code.

Let me know if there is any additional details that you think would merit clarification.
