### 2025-07-23 #673 @tataphani
Hello Ropensci/Software-Review ,

I agree with you that installation is biggest hurdle , My concern was that
I am taking too much of time for turnaround because of day job and other
commitments , I am taking longer than expected time very sorry for that .
For that reason only I am little bit shy and guilty too . I will give it
try again on posit online ( I forgot about that option ) to see if that
works .

Very sorry for the delay

thanks
Phani

On Wed, Jul 23, 2025 at 8:46 AM Mauro Lepore ***@***.***>
wrote:

> *maurolepore* left a comment (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673#issuecomment-3108639554>
>
> Thanks so much @tataphani <https://github.com/tataphani> for your effort!
> Your installation struggles are in themselves useful feedback.
>
> As a developer it's useful to know what's stopping users of your software,
> and installation issues is the fist and often biggest barrier. Often the
> issue is not the software but the environment. The developer has little
> control on that but it's still useful to know what those environments look
> like, and how to nudge users towards environments with less friction.
>
> For example, it's possible to pre-install software in a service like
> https://login.posit.cloud/ and share the entire project with the user.
> Maybe @ryanzomorrodi <https://github.com/ryanzomorrodi> and you tried
> this out already?
>
> If you still have a lil time and energy left, to make the most of the
> effort you already made it would be wonderful to have a comment summarizing
> the things you tried and the problems you faced, ideally pasting some code
> and error messages but fine if that's already gone. That would help me and
> @ryanzomorrodi <https://github.com/ryanzomorrodi> be more effective in
> supporting future reviewers of rOpenSci, and future users of {sasquatch}.
>
> Totally fine if not, and in any case thanks again!
>
