### 2025-07-23 #673 @maurolepore
> Upon changes being made, change the review status tag to 5/awaiting-reviewer-response, and request that reviewers indicate approval with the [reviewer approval template](https://devdevguide.netlify.app/approval2template). -- https://devdevguide.netlify.app/softwarereview_editor

---

@thisisnic could you please indicate if you would request further changes? If not, please indicate your approval using [Reviewer approval comment template](https://devdevguide.netlify.app/approval2template).

I'll the same ASAP, based on the two rounds of feedback I provided during my editor checks.
