### 2025-07-23 #673 @tataphani
Thanks, I'll check it out!

On Wed, Jul 23, 2025 at 3:25 PM Ryan Zomorrodi ***@***.***>
wrote:

> *ryanzomorrodi* left a comment (ropensci/software-review#673)
> <https://github.com/ropensci/software-review/issues/673#issuecomment-3110050489>
>
> @tataphani <https://github.com/tataphani>
>
> You can use this posit cloud project: https://posit.cloud/content/10632695
>
> I think you use US 1 as your SAS server, so I set it up for that. If not,
> just run sas_configure("oda", overwrite = TRUE) and answer with whatever
> server you use.
>
> You will need to add your username and password to
> /cloud/home/r1261695/.authinfo. It should be already open when you open
> the project. You just need to replace the {username} and {password}:
>
> oda user {username} password {password}
>
