### 2025-07-28 #673 @maurolepore
@ryanzomorrodi

I'm reviewing your responses to my comments and overall I'm excited about all the work you've done. Congratulations!

Here are some comments. As before, I expect some response from the ones marked with checkboxes. The rest are optional.

- [ ] ml32. The major one is that the package doesn't pass our automated checks. The the comment above.
> ✖️ Package fails continuous integration checks.
> ✖️ Package coverage is 26.6% (should be at least 75%)
> Important: All failing checks above must be addressed prior to proceeding

- [ ] ml33. On closer inspection I noticed the .github/ directory lacks workflows to run R CMD check on GitHub actions. Please see the [CI section of the dev-guide](https://devdevguide.netlify.app/pkg_ci) and `[?usethis::use_github_actions("check-standard")](https://usethis.r-lib.org/reference/github_actions.html)`.

- [ ] ml34. A good way to ensure the package is structurally solid is to walk through the standard checks for CRAN submission. Even if you don't submit to CRAN, it's a good exercise. It can reveal issues in a pretty automated way. See `[?usethis::use_release_issue()](https://usethis.r-lib.org/reference/use_release_issue.html)`.

- [ ] ml35. The tests directory contains not only unit tests but also integration tests. Evidence of this is that running `devtools::test()` in the posit-cloud project doesn't "just work".

```r
> devtools::test()
ℹ Testing sasquatch
✔ | F W  S  OK | Context
✖ | 1        1 | configure [3.6s]
────────────────────────────────────────────────────────────────────────────────────
Error (test-configure.R:46:3): oda configuration - authinfo
Error in `.rs.api.documentOpen(filePath, line = line, col = col, moveCursor = moveCursor)`: /home/r68453/.authinfo does not exist.
Backtrace:
    ▆
 1. ├─base::suppressMessages(write_authinfo_template(overwrite = TRUE)) at test-configure.R:46:3
 2. │ └─base::withCallingHandlers(...)
 3. └─sasquatch:::write_authinfo_template(overwrite = TRUE)
 4.   └─rstudioapi::navigateToFile(authinfo_path) at sasquatch/R/saspy-configure.R:178:5
 5.     └─rstudioapi::callFun(...)
 6.       ├─base::do.call(f, args)
 7.       └─`<fn>`("/home/r68453/.authinfo", -1L, -1L, TRUE)
 8.         └─.rs.api.documentOpen(filePath, line = line, col = col, moveCursor = moveCursor)
────────────────────────────────────────────────────────────────────────────────────
⠋ |          1 | connect

Please enter the OMR user id:
```

Maybe it's too hard to make the tests fully independent from the ODA server but we need to figure out a way for our CI systems to be able to run without failure. One way to solve that problem might be to configure the CI system with the necessary environment variables to access a test-environment. Many of our packages need to access external resources (e.g. APIs), so you may explore how they do it.

I think we already discussed this issue and I know it's a really hard one :-(. The relevant section of the devguide is [Testing](https://devdevguide.netlify.app/pkg_building#testing).

> If your package interacts with web resources (web APIs and other sources of data on the web) you might find the [HTTP testing in R book by Scott Chamberlain and Maëlle Salmon](https://books.ropensci.org/http-testing/) relevant. Packages helping with HTTP testing (corresponding HTTP clients):

- [httptest2](https://enpiar.com/httptest2/) ([httr2](https://httr2.r-lib.org/));
- [httptest](https://enpiar.com/r/httptest/) ([httr](https://httr.r-lib.org/));
- [vcr](https://docs.ropensci.org/vcr/) ([httr](https://httr.r-lib.org/), [crul](https://docs.ropensci.org/crul));
- [webfakes](https://webfakes.r-lib.org/) ([httr](https://httr.r-lib.org/), [httr2](https://httr2.r-lib.org/), [crul](https://docs.ropensci.org/crul), [curl](https://jeroen.r-universe.dev/curl#)).

BTW, after configuring ODA the tests work fine.

```
> sasquatch::configure_saspy(template = "oda", overwrite = T)
Which server is your account on?

1: United States 1
2: United States 2
3: Europe 1
4: Asia Pacific 1
5: Asia Pacific 2

Selection: 1
✔ Writing to /cloud/project/r-saspy/lib/python3.12/site-packages/saspy/sascfg_personal.py.
✔ Writing to /home/r68453/.authinfo.
ℹ Replace "{username}" and "{password}" with your ODA username and password within
  /home/r68453/.authinfo.
ℹ Download and extract the SAS ODA zip encryption jars
  (<https://support.sas.com/downloads/package.htm?pid=2494>) into
  /cloud/project/r-saspy/lib/python3.12/site-packages/saspy/java/iomclient/.
ℹ For more information about SASPy configuration see the sasquatch::configuration
  vignette or the SASPy documentation.
> library(sasquatch)
> sas_connect()
✔ SAS connection established.
> devtools::test()
ℹ Testing sasquatch
✔ | F W  S  OK | Context
✔ |          2 | configure [1.2s]
✔ |         10 | connect [33.6s]
✔ |          7 | connection-checks [22.6s]
✔ |          3 | disconnect [1.9s]
✔ |         15 | engine [11.7s]
✔ |          5 | file-copy [6.4s]
✔ |          7 | file-download [8.9s]
✔ |          2 | file-exists [5.0s]
✔ |          4 | file-remove [5.0s]
✔ |          3 | file-upload [4.5s]
✔ |         27 | from-r [1.4s]
✔ |          4 | list [1.7s]
✔ |          3 | run-file [3.6s]
✔ |          1 | run-string [1.1s]
✔ |         17 | to-r [80.2s]

══ Results ═════════════════════════════════════════════════════════════════════════
Duration: 188.7 s

[ FAIL 0 | WARN 0 | SKIP 0 | PASS 110 ]
```

- [ ] ml36. On posit claud R CMD check passes cleanly but this message suggests the need for an explicit mention of the minimal R version it depends on. Once you identified such minimal version, it would be great to run R CMD check on CI with that version and a few more recent ones, for all three main platforms.

```r
     NB: this package now depends on R (>= 3.5.0)
     WARNING: Added dependency on R >= 3.5.0 because serialized objects in
     serialize/load version 3 cannot be read in older versions of R.
     File(s) containing such objects:
       ‘sasquatch/vignettes/files/proc_means_widget.rds’
     NB: this package now depends on R (>= 4.1.0)
     WARNING: Added dependency on R >= 4.1.0 because package code uses the
     pipe |> or function shorthand \(...) syntax added in R 4.1.0.
     File(s) using such syntax:
       ‘engine.R’ ‘from-r.R’ ‘to-r.R’
─  building ‘sasquatch_0.0.0.9043.tar.gz’
   Warning: invalid uid value replaced by that for user 'nobody'
```

- ml37. Just the posit cloud project you shared worked great, but locally I failed to install all the requirements. Any idea of what might be going on? (I did update `pyenv` and I do have `python3`. My credentials work great on the posit cloud project.)

```r
> reticulate::install_python()
+ /opt/homebrew/bin/pyenv update
+ /opt/homebrew/bin/pyenv install --skip-existing NA
python-build: definition not found: NA

See all available versions with `pyenv install --list'.

If the version you need is missing, try upgrading pyenv:

  brew update && brew upgrade pyenv
Error: installation of Python NA failed

> system("which python")
> system("which python3")
/Library/Frameworks/Python.framework/Versions/3.13/bin/python3
> sasquatch::sas_connect()
Error in py_call_impl(callable, call_args$unnamed, call_args$named) :
  saspy.sasexceptions.SASIOConnectionTerminated: No SAS process attached. SAS process has terminated unexpectedly.
<class 'Exception'>
Run `reticulate::py_last_error()` for details.
```

My system

```
 Sys.info()
                                                                                                 sysname
                                                                                                "Darwin"
                                                                                                 release
                                                                                                "24.5.0"
                                                                                                 version
"Darwin Kernel Version 24.5.0: Tue Apr 22 19:54:43 PDT 2025; root:xnu-11417.121.6~2/RELEASE_ARM64_T8132"
                                                                                                nodename
                                                                              "Mauros-MacBook-Pro.local"
                                                                                                 machine
                                                                                                 "arm64"
                                                                                                   login
                                                                                                  "root"
                                                                                                    user
                                                                                           "maurolepore"
                                                                                          effective_user
                                                                                           "maurolepore"
```

- ml38. NEWS might be best reserved for communication with the wider community. For communications with reviewers this thread seems like the best (and official) place.
- ml39. README might benefit from some text-based example, that can be copied and pasted rather than images only).
- ml40. README might benefit from moving the comparison with other packages to the end.
