### 2025-07-28 fc58048
Add min R version
