### 2025-07-31 37229cc
Add CI back in
