### 2025-07-31 afb1369
Reduce down the NEWS.md
