+++ NEWS.md
@@ -2,71 +2,8 @@
 
 ## rOpenSci Review Changes
 
-- `configure_saspy()` no longer asks that users type in their SAS ODA username and password within the terminal. See [secrets vignette](https://cran.r-project.org/web/packages/httr/vignettes/secrets.html).  
-  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):
-
-  > I did feel a little uncomfortable entering in my SAS online password into the command line - I would have preferred to be given information about the config file and where it's stored to enter it there myself as that seems more typical of how other packages handle this kind of thing.
-  > 
-  > The password remains in the R terminal unobscured after being entered, which isn't great security practice.
-
+- `configure_saspy()` no longer asks that users type in their SAS ODA username and password within the terminal. See `vignette("secrets", "httr")`.  
 - Dataset used to show `sas_to_r()` in `README.md` changed to `warpbreaks` and the variable name of the version of `"sashelp.cars"` was renamed to `sas_cars`.
-  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):  
-
-  > I tried another example from the README
-  >
-  > sas_from_r(mtcars, "mtcars")
-  > Warning message:
-  > In sas_from_r(mtcars, "mtcars") :
-  > x rownames will not be transferred as a column.
-  >
-  > It took me a moment to understand that x refers to the source table - this could be rephrased to be clearer, or a different example used which doesn't result in a warning message.
-  > It'd also be handy to have it explained explicitly that this creates it as "mtcars" in the current SAS session - this wasn't clear to me and I had to check the docs to find this out.
-  >
-  > cars <- sas_to_r("cars", libref = "sashelp")
-  > 
-  > As there is already a dataset in R called cars I might be tempted to call this variable something else.
-    
 - Fixed typo in `configure_saspy()`. Changed Europe 2 to Europe 1.  
-  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):
-
-  > When I went to configure it, the setup script asked "Which server is your account on?" with the following options:
-  > 
-  > 1. United States 1
-  > 2. United States 2
-  > 3. Europe 2
-  > 4. Asia Pacific 1
-  > 5. Asia Pacific 2
-  > 
-  > Mine is "Europe 1" but I was unable to select this. This didn't prevent it from running in the end though.
-
 - Specified `rlang` imports.  
-  From [thisisnic](https://github.com/ropensci/software-review/issues/673#issuecomment-2833364073):
-
-  > Running devtools::check()  
-  >   
-  > \u276f checking package subdirectories ... NOTE  
-  >   Problems with news in 'NEWS.md':  
-  >   No news entries found.  
-  >   
-  > \u276f checking R code for possible problems ... NOTE  
-  >   .rlang_check_is_string: no visible binding for global variable 'na_chr'  
-  >   .rlang_stop_unexpected_typeof: no visible global function definition  
-  >     for 'caller_env'  
-  >   check_logical: no visible global function definition for 'caller_arg'  
-  >   check_logical: no visible global function definition for 'caller_env'  
-  >   stop_input_type: no visible global function definition for 'caller_arg'  
-  >   stop_input_type: no visible global function definition for 'caller_env'  
-  >   Undefined global functions or variables:  
-  >     caller_arg caller_env na_chr  
-  >   
-  > Looking at the second of those, it looks like the notes were generated by the signature for check_logical()  
-  >   
-  > https://github.com/ryanzomorrodi/sasquatch/blob/18cd7ca03a107be78700fdd3febf6908f28563c6/R/import-standalone-types-check.R#L508-L509  
-  >   
[…]
