### 2025-07-31 14d4087
Move comparisons to bottom and add text from within the pictures
