### 2025-08-01 28494ec
Add additional clarification if saspy cannot be found
