### 2025-08-02 4037bc3
Make install saspy more flexible
