### 2025-08-02 #673 @ryanzomorrodi
Just to follow up on this within this thread. It seems like rstudio has a setting which calls `reticulate::use_python("some python path")` so after you use `install_saspy()` which installs saspy in a separate virtual environment, saspy is still not found by any other function.

I added some additional notes about this problem within the documentation and you now get this error:

```r
reticulate::use_python("/usr/bin/python3")
sasquatch::sas_connect()
#> ✖ No SASPy installation found within /usr/bin/python3.
#> ℹ Python version was forced by use_python() function.
#> ℹ If SASPy is installed elsewhere, set the Python version with `reticulate::use_python()`, `reticulate::use_virtualenv()`, `reticulate::use_conda()`, or `reticulate::use_miniconda()`.
#> ℹ Else, use `sasquatch::install_saspy()` and `reticulate::use_virtualenv()` to install SASPy within a virtual env and set it as your Python version.
```

Which is at very least more informative before, even though it may be confusing for users that it says `use_python()` was used when they never used it directly. I'm open to other ideas, but I think this maintains the ability for users to use whatever python version they want which I think is pretty important. I also don't know how prevalent having this setting on is.

Last thing, I made `install_saspy()` able to install conda environments as well to kind of fit this ethos a little bit better.
