### 2025-08-03 f15a361
skip sas_connect test if no jars present
