### 2025-08-03 #673 @ryanzomorrodi
ml32. I already have quite a bit of `skip_if_*` statements, which should skip tests in most cases where the developer has not set up their saspy configuration yet.

So for example, the CI should not fail in the ropensci CI system because it does not have saspy installed nor does it have it configured. This is from my computer but you can see if I remove the `saspy` virtual env, all the integration tests are skipped.

```r
devtools::load_all()
get_saspy_path()
/home/user/.virutalenvs/r-saspy/lib/python3.12/site-packages/saspy/
```

```bash
rm -rf ~/.virutalenvs/r-saspy
```

```r
devtools::test()
#> ℹ Testing sasquatch
#> ✔ | F W  S  OK | Context
#> ✔ |      2   0 | configure
#> ✔ |      5   0 | connect [1.5s]
#> ✔ |      5   1 | connection-checks [1.5s]
#> ✔ |      2   0 | disconnect
#> ✔ |      3   9 | engine
#> ✔ |      2   0 | file-copy
#> ✔ |      3   0 | file-download
#> ✔ |      2   0 | file-exists
#> ✔ |      2   0 | file-remove
#> ✔ |      2   0 | file-upload
#> ✔ |         27 | from-r
#> ✔ |      2   0 | list
#> ✔ |      3   0 | run-file
#> ✔ |      1   0 | run-string
#> ✔ |      9   0 | to-r [2.7s]
#>
#> ══ Results ═══════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════════
#> Duration: 13.6 s

#> ── Skipped tests (43) ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────
#> • SASPy is not installed. (43): test-configure.R:3:3, test-configure.R:40:3, test-connect.R:4:3, test-connect.R:28:3, test-connect.R:46:3, test-connect.R:66:3, test-connect.R:89:3, test-connection-checks.R:4:3,
#>   test-connection-checks.R:13:3, test-connection-checks.R:22:3, test-connection-checks.R:37:3, test-connection-checks.R:52:3, test-disconnect.R:4:3, test-disconnect.R:14:3, test-engine.R:4:3, test-engine.R:22:3,
#>   test-engine.R:54:3, test-file-copy.R:4:3, test-file-copy.R:19:3, test-file-download.R:4:3, test-file-download.R:23:3, test-file-download.R:38:3, test-file-exists.R:4:3, test-file-exists.R:15:3, test-file-remove.R:4:3,
#>   test-file-remove.R:19:3, test-file-upload.R:4:3, test-file-upload.R:21:3, test-list.R:4:3, test-list.R:16:3, test-run-file.R:4:3, test-run-file.R:19:3, test-run-file.R:40:3, test-run-string.R:4:3, test-to-r.R:4:3,
#>   test-to-r.R:24:3, test-to-r.R:48:3, test-to-r.R:72:3, test-to-r.R:94:3, test-to-r.R:121:3, test-to-r.R:143:3, test-to-r.R:171:3, test-to-r.R:197:3
#>
#> [ FAIL 0 | WARN 0 | SKIP 43 | PASS 37 ]
```

The only way to get around the skip statements, is if you install saspy, then configure it incorrectly (unlikely because sas_configure("oda") does all nearly all the work and I just added a check that makes it skip if you don't have the encryption jars OR you typed in your password wrong in `.authfile`

As you can tell nearly all the tests are integration tests because fundamentally sasquatch is a a convenience wrapper around SASPy.
