### 2025-08-03 18246ff
specify type in vapply within sas_to_r
