### 2025-08-16 #673 @ropensci-review-bot
Approved! Thanks @ryanzomorrodi for submitting and @thisisnic, @tataphani for your reviews! :grin:
